% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adegenet_DT2genind.R
\name{adegenet_DT2genind}
\alias{adegenet_DT2genind}
\title{Convert a long genotype data table into a \code{genind} object}
\usage{
adegenet_DT2genind(dat, sampCol = "SAMPLE", locusCol = "LOCUS",
  genoCol = "GT", popCol = NULL, n = 2L)
}
\arguments{
\item{dat}{Data.table: Contains samples and loci in rows, with a
separate column endcoding genotypes, where each allele is separated
by a '/'.}

\item{sampCol}{Character: Column with sample information. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: Column with locus information. Default = \code{'LOCUS'}.}

\item{genoCol}{Character: Column with genotype information. Default = \code{'GT'}.}

\item{popCol}{Character: Column with population information. Optional.}

\item{n}{Integer: The ploidy level. Default = 2.}
}
\value{
A \code{genind} object, with the slot \code{pop} slot filled if argumnet
\code{popCol} is specified.
}
\description{
For a data table with samples and loci as rows (long-format data table),
create a \code{genind} object, as per the \code{adegenet} package
(Jombart 2008 Bioinformatics). Does not have to be biallelic.
}
\examples{
data(genomalicious_4pops)

genomalicious_4pops

DT2genind(genomalicious_4pops)
DT2genind(genomalicious_4pops, popCol='POP')

}
\references{
Jombart (2008) adegenet: a R package for the multivariate analysis of genetic markers. Bioinformatics.
}
