% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adegenet_DT2genlight.R
\name{adegenet_DT2genlight}
\alias{adegenet_DT2genlight}
\title{Convert a long genotype data table into a \code{genlight} object}
\usage{
adegenet_DT2genlight(dat, sampCol = "SAMPLE", locusCol = "LOCUS",
  genoCol = "GT", popCol = NULL)
}
\arguments{
\item{dat}{Data.table: Contains samples and loci in rows, with a
separate column endcoding genotypes, where each allele is separated
by a '/', or as a count of ref alleles (0, 1 or 2). Assumes biallelic loci.}

\item{sampCol}{Character: Column with sample information. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: Column with locus information. Default = \code{'LOCUS'}.}

\item{genoCol}{Character: Column with genotype information. Default = \code{'GT'}.}

\item{popCol}{Character: Column with population information. Optional.}
}
\value{
A \code{genind} object, with the slot \code{pop} slot filled if argumnet
\code{popCol} is specified.
}
\description{
For a data table with samples and loci as rows (long-format data table),
create a \code{genlight} object, as per the \code{adegenet} package
(Jombart 2011 Bioinformatics). Must be biallelic.
}
\examples{
data(genomalicious_4pops)

genomalicious_4pops

adegenet_DT2genlight(genomalicious_4pops)
adegenet_DT2genlight(genomalicious_4pops, popCol='POP')

}
\references{
Jombart (2008) adegenet 1.3-1: new tools for the analysis of genome-wide SNP data. Bioinformatics.
}
