% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_mat_rm_gaps.R
\name{align_mat_rm_gaps}
\alias{align_mat_rm_gaps}
\title{Remove gaps from an alignment matrix}
\usage{
align_mat_rm_gaps(align)
}
\arguments{
\item{align}{Matrix: A sequence alignment converted into a matrix, with
positions in columns, samples in rows, and cells being DNA nucleotides or
amino acids. Rowsnames should be names of samples.}
}
\value{
Returns a matrix with same number of rows as `align`, but with any
columns with `'-'` being removed. Columns are relabelled as `'pos.X'`, with
'X', being the relative position in the original alignment.
}
\description{
Provide a sequence alignment in matrix format and remove gaps ('-').
}
