% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_plot_dna.R
\name{align_plot_dna}
\alias{align_plot_dna}
\title{Plot DNA alignments}
\usage{
align_plot_dna(
  alignMat,
  nuc_colours = NULL,
  other_colours = NULL,
  gap_colour = NA,
  border_colour = NA,
  pos_vec = NULL,
  show_bases = TRUE,
  base_size = 3,
  show_legend = FALSE
)
}
\arguments{
\item{alignMat}{Matrix: the DNA alignments. Rows are the taxa, columns are
the base positions, and cells contain the values. Values should be one of
'A', 'T', 'G', 'C', '?', 'N', '-'.}

\item{nuc_colours}{Character: named vector of colour values for nucleotides.
For example, \code{c('A'='blue', 'T'='yellow', 'G'='green', 'C'='red')}.
Default is NA, which will produce automated colours.}

\item{other_colours}{Character: named vector colour values for '?' or 'N'
values. For example, \code{c('?'='black', 'N'='black')}.
Default is NA, which will produce automated colours.}

\item{gap_colour}{Chracter: named vector colour for gaps. For examples,
\code{c('-'='white')}. #' Default is NA, which will produce
automated colours.}

\item{border_colour}{Character: single value, the colour of borders around
base positions. Default is NA, no assigned border colour.}

\item{pos_vec}{Integer: a vector of base positions to plot, subsets alignment.
Default is NULL, which will plot all bases.}

\item{show_bases}{Logical: whether the values of bases should be plotted.
Default is TRUE.}

\item{base_size}{Numeric: the text size used for nucleotide bases. Default is 3.}

\item{show_legend}{Logical: whether a legend should be plotted.
Default is FALSE.}
}
\description{
This function takes in an alignment of DNA sequences in a matrix format
and returns a plot of that alignment.
}
\examples{
library(genomalicious)

# Create a link to raw external datasets in genomalicious
genomaliciousExtData <- paste0(find.package('genomalicious'), '/extdata')

# Path to the demo FASTA file
fasta <- paste0(genomaliciousExtData, '/data_COI.fasta')

# Multi sequence alignmnet of demo COI data.
aln <- align_many_genes_dna(fasta, gene.names='COI')

# Plot base positions from 200:250
align_plot_dna(as.matrix(aln$COI$align), pos_vec=200:250)

# Custom colours, no text
align_plot_dna(
   as.matrix(aln$COI$align),
   nuc_colours=c(`A`='#ce0073',`T`='#e46adf',`G`='royalblue',`C`='#08c7e0'),
   pos_vec=200:250,
   show_bases=FALSE
)

}
