% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rm_gaps_invars.R
\name{align_rm_gaps_invars}
\alias{align_rm_gaps_invars}
\title{Remove gaps and invariant sites from an alignment matrix}
\usage{
align_rm_gaps_invars(alignMat, removeGaps, removeInvars)
}
\arguments{
\item{alignMat}{Matrix: A sequence alignment converted into a matrix, with
positions in columns, samples in rows, and cells being DNA nucleotides or
amino acids. Row names should be names of samples.}

\item{removeGaps}{Logical: Should gaps be removed? `TRUE` or `FALSE`.}

\item{removeInvars}{Logical: Should invariant sites be removed? `TRUE` or `FALSE`.}
}
\value{
Returns a matrix with same number of rows as `alignMat`, but with any
columns removed that are gaps and/or invariant sites.
Columns are relabeled as `'pos.X'`, with 'X', being the relative
position in the original alignment.
}
\description{
Provide a sequence alignment in matrix format and remove gaps and/or
invariant sites.
}
\details{
Gaps will be removed before invariant sites. Removal of gaps
occurs for any columns in `alignMat` that possess a `'_'`. Removal of invariant
sites occurs for any column in `alignMat` where `length(unique(x))` does
not evaluate to >2. Note, this will also include gaps if you do not remove them.
}
