% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_freq_corr_windows.R
\name{allele_freq_corr_windows}
\alias{allele_freq_corr_windows}
\title{Allele frequency correlations across genomic windows}
\usage{
allele_freq_corr_windows(
  dat,
  chromCol = "CHROM",
  posCol = "POS",
  locusCol = "LOCUS",
  popCol = "POP",
  freqCol = "FREQ",
  windowSize = 10000,
  numLoci = 0
)
}
\arguments{
\item{dat}{Data table: Contains populations allele frequencies.}

\item{chromCol}{Character: The column with chromosome information.
Default is 'CHROM'.}

\item{posCol}{Character: The column with position information.
Default is 'POS'.}

\item{locusCol}{Character: The column with locus ID.
Default is 'LOCUS'.}

\item{popCol}{Character: The column with population ID.
Default is 'POP'.}

\item{freqCol}{Character: The column with frequencies of a focal allele.
Default is 'FREQ'.}

\item{windowSize}{Integer: The window size to use. Default is 10000.}

\item{numLoci}{Integer: The number of loci to subsample per window. You may
need to adjust this value depending on the density of markers in your dataset.
Default is 0, which triggers all loci to be used. Specify >0 if you want
to control the subsampling.}
}
\value{
Returns a long-format data table of pairwise correlations between
loci within windows, within chromosomes. Has the columns:
\enumterate{
 \item \code{$CHROM}, the chromosome ID.
 \item \code{$WINDOW}, the window ID.
 \item \code{$LOCUS.1}, the locus 1 ID.
 \item \code{$CHROM}, the locus 2 ID.
 \item \code{$DIST}, the distance between loci in base pairs.
 \item \code{$CORR}, the Pearson correlation between locus allele frequencies.
}
}
\description{
This function takes a data table of population allele frequencies
across different chromosomes/contigs and estimates correlations between
loci within specified windows. Loci within these windows can be
randomly subsmapled to reduce computational time and memory cost.
}
\examples{
library(genomalicious)

# RADseq pool-seq dataset
data(data_PoolFreqs)

# Note, only subset of contigs have multiple alleles, so only
# a few pairwise loci are returned.
allele_freq_corr_windows(data_PoolFreqs)
}
