% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_uniq_DT.R
\name{allele_uniq_DT}
\alias{allele_uniq_DT}
\title{Generate a data table of unique alleles per locus}
\usage{
allele_uniq_DT(
  dat,
  chromCol = "CHROM",
  posCol = "POS",
  locusCol = "LOCUS",
  refCol = "REF",
  altCol = "ALT"
)
}
\arguments{
\item{dat}{Data.table: Long-format data table of variants, e.g., as read in
with \code{genomalicious::vcf2DT}.}

\item{chromCol}{Character: The column with chromosome information. Default is "CHROM".}

\item{posCol}{Character: The column with position information. Default is "POS".}

\item{locusCol}{Character: The column with locus ID information. Default is "LOCUS".}

\item{refCol}{Character: The column with reference allele information. Default is "REF".}

\item{altCol}{Character: The column with alternate allele information. Default is "ALT".}
}
\value{
Returns a wide-format data table with columns \code{$CHROM}, \code{$POS},
and \code{$LOCUS}, and multiple \code{$ALLELE.X} columns, where 'X' is the
i-th allele at a locus. The maximum number of allele columns will depend on
whichever locus has the most alleles. Columns with NA indicate absence of an
i-th allele, otherwise, the allele sequence will fill the cell. The reference
allele is \code{$ALLELE.0}, with subsequent columns representing alternate
alleles, e.g., \code{$ALLELE.1}, \code{$ALLELE.2}, etc.
}
\description{
Takes a data.table of variants in long-format and converts to a wide-format
data.table of unique alleles. Can be used for multiallelic datasets.
}
\examples{

library(genomalicious)

data(data_Genos)

allele_uniq_DT(data_Genos)

}
