% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleReadsDT.R
\name{alleleReadsDT}
\alias{alleleReadsDT}
\title{Generate an data.table of allelic read depth}
\usage{
alleleReadsDT(
  dat,
  chromCol = "CHROM",
  posCol = "POS",
  locusCol = "LOCUS",
  sampCol = "SAMPLE",
  refCol = "REF",
  altCol = "ALT",
  roCol = "RO",
  aoCol = "AO"
)
}
\arguments{
\item{dat}{Data.table: The input data of variants, samples, and read counts.
Requires the following columns:
\enumerate{
  \item The chromosome ID (see param \code{chromCol}).
  \item The variant position (see param \code{posCol}).
  \item The locus ID (see param \code{locusCol}).
  \item The sampled individuals (see param \code{sampCol}).
  \item The reference allele (see param \code{refCol}).
  \item The alternate allele(s) (see param \code{altCol}).
  \item The reference read observations (see param \code{roCol}).
  \item The alternate read observations (see param \code{aoCol}).
}}

\item{chromCol}{The column with chromosome information. Default is \code{'CHROM'}.}

\item{posCol}{The column with variant position information. Default is \code{'POS'}.}

\item{locusCol}{The column with locus ID information. Default is \code{'LOCUS'}.}

\item{sampCol}{The column with sample ID information. Default is \code{'SAMPLE'}.}

\item{refCol}{The column with Ref allele information. Default is \code{'REF'}.}

\item{altCol}{The column with Alt information. For loci with more than one
Alt allele, each allele's nucleotides should be separated by a comma. Default is \code{'ALT'}.}

\item{roCol}{The column with Ref read observations. Default is \code{'RO'}.}

\item{aoCol}{The column with Alt read observations. For loci with more than one
Alt allele, each allele's reads should be separated by a comma. Default is \code{'CHROM'}.}
}
\value{
Returns a long format data.table with the columns:
\enumerate{
   \item \code{$CHROM}: character, the chromosome ID.
   \item \code{$POS}: integer, the variant position ID.
   \item \code{$LOCUS}: character, the locus ID.
   \item \code{$SAMPLE}: character, the sample ID.
   \item \code{$ALLELE}: character, the allele nucleotides.
   \item \code{$IS.REF}: logical, is the allele the Ref allele?
   \item \code{$READS}: integer, the observed reads for the allele.
}
}
\description{
A data.table of samples and read observations in RO (Ref) and AO (Alt) columns
is converted into another long-format data.table of counts for each allele.
This stacks Ref and Alt alleles into a single column and can be used on loci
with >2 alleles.
}
\examples{
library(genomalicious)
data("data_Genos")

readsPerAllele <- alleleReadsDT(data_Genos)
readsPerAllele

}
