% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_correct_gaps.R
\name{assembly_correct_gaps}
\alias{assembly_correct_gaps}
\title{Correct uneven gaps in an assembled genome}
\usage{
assembly_correct_gaps(genomeSS, threshold, correct)
}
\arguments{
\item{genomeSS}{DNAStringSet: the genome assembly as a 'DNAStringSet'
object from the \code{Biostrings} package.}

\item{threshold}{Integer: the minimum string of 'N's to consider a gap.}

\item{correct}{Integer: the standardised number of 'N's for gaps of unknown size.}
}
\value{
Returns the genome assembly back as a DNAStringSet object with all
gaps meeting the threshold size standardized to the corrected size.
Sequences with corrected gaps are labelled with '_correct_gaps'.
}
\description{
Uneven gap sizes can form during genome assembly as different
assembling, polishing, and gap closing tools pile on top of each
other. This makes submission of a genome assembly problematic,
because varying gap sizes are ambiguous, and ideally we want unknown
gaps to have a standardised size. \cr\cr
This function takes a genome assembly and corrects gap sizes given
a user-define threshold. It is assumed that any gap larger than the
threshold is an unknown gap.
}
