% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayescan_inputs.R
\name{bayescan_inputs}
\alias{bayescan_inputs}
\title{Generate \code{bayescan} input files}
\usage{
bayescan_inputs(dat, pool.info, file.bayescan, file.loci)
}
\arguments{
\item{dat}{Data.table: The biallelic SNP data. Requires all of the following columns: \cr
\enumerate{
  \item \code{POOL} = The population pool ID. \cr
  \item \code{LOCUS} = The locus ID. \cr
  \item \code{PI} = The population Ref allele frequency. \cr
}}

\item{pool.info}{Data table: The population pool metadata. Requires all of the following columns: \cr
\enumerate{
  \item \code{$POOL} = The population pool ID. \cr
  \item \code{$INDS} = The number of diploid individuals in each pooled library.
}}

\item{file.bayescan}{Character: The name of the input file for \code{bayescan}/}

\item{file.loci}{Character: A file that contains an information about the locus ID in
the \code{bayescan} input file (i.e., \code{file.bayescan}).}
}
\description{
Generates an input file for Bayescan from pooled allele frequencies. \cr
\cr
It assumes that these pooled allele frequencies have been estimated from Gautier et al. (2013)'s
\code{poolne_estim}, and have been imported into \code{R} with \code{pgposer}.
}
\details{
The allele counts in the Bayescan input file generated reflect the number of haploid
genomes pooled. E.g. if 20 individuals were pooled, i.e. 40 genomes, and the Ref allele was
estimated at a frequency of, the counts would be Ref=28 and Alt=12.
}
\examples{
data(pgposerPi)
data(pgposerInfo)

Bayescan_gen_inputs(pgposerPi, pgposerInfo, 'Bayescan_input.txt', 'Bayescan_loci.txt')

}
