% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescan_inputs_pool.R
\name{bayescan_inputs_pool}
\alias{bayescan_inputs_pool}
\title{Generate \code{bayescan} input files}
\usage{
bayescan_inputs_pool(
  dat,
  pool.info,
  file.bayescan,
  file.loci,
  poolCol,
  locusCol,
  freqCol
)
}
\arguments{
\item{dat}{Data table: The biallelic SNP data. Requires all of the following columns: \cr
\enumerate{
  \item The population pool ID. \cr
  \item The locus ID. \cr
  \item The population Ref allele frequency. \cr
}}

\item{pool.info}{Data table: The population pool metadata. Requires all of the following columns: \cr
\enumerate{
  \item \code{$POOL} = The population pool ID. \cr
  \item \code{$INDS} = The number of diploid individuals in each pooled library.
}}

\item{file.bayescan}{Character: The name of the input file for \code{bayescan}/}

\item{file.loci}{Character: A file that contains an information about the locus ID in
the \code{bayescan} input file (i.e., \code{file.bayescan}).}

\item{poolCol}{Character: Population pool ID. Default = \code{'POOL'}}

\item{locusCol}{Character: Locus ID. Default = \code{'LOCUS'}}

\item{freqCol}{Character: The reference allele frequency. Default = \code{'FREQ'}.}
}
\description{
Generates an input file for Bayescan (Foll & Gaggiotti, 2008) from pooled allele frequencies. \cr
\cr
It assumes that these pooled allele frequencies have been estimated from Gautier et al. (2013)'s
\code{poolne_estim}, and have been imported into \code{R} with \code{genomalicious}.
}
\details{
The allele counts in the Bayescan input file generated reflect the number of haploid
genomes pooled. E.g. if 20 individuals were pooled, i.e. 40 genomes, and the Ref allele was
estimated at a frequency of 0.7, the counts would be Ref=28 and Alt=12. NOTE: The estimated number
of indivdiuals carrying an allele is rounded to nearest integer (e.g. 1.5 = 2, and 1.4 = 1),
with the exception when the number of individauls is < 1 but > 0, in which it is always rounded to 1.
}
\examples{
data(data_PoolPi)
data(data_PoolInfo)

bayescan_inputs_pool(dat=data_PoolPi
                     , pool.info=data_PoolInfo
                     , file.bayescan='Bayescan_input.txt'
                     , file.loci='Bayescan_loci.txt'
                     , poolCol='POOL', locusCol='LOCUS', freqCol='PI')

}
\references{
Foll & Gaggiotti (2008) A genome scan method to identify selected loci appropriate
for both dominant and codominant markers: A Bayesian perspective. Genetics 180: 977-993.
}
