% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2featuretab.R
\name{bed2featuretab}
\alias{bed2featuretab}
\title{Converts a BED file of annotations into a GenBank feature table.}
\usage{
bed2featuretab(bedFiles, featureRef, outFile)
}
\arguments{
\item{bedFiles}{Character: A vector of BED file names to process.}

\item{featureRef}{Data.table: This contains the list of features that are to
be mapped against the annotations in BED format. Requires 3 columns:
\enumerate{
   \item \code{$NAME}, the name of the feature, which matches the 4th column
   of the BED file.
   \item \code{$FEATURE}, the type of feature, e.g., 'gene', 'tRNA', or 'misc_feature'.
   \item \code{$DETAIL}, the details to accompany the feature, which could be
   the gene or product name, or notes to attach to misc_features.
}}

\item{outFile}{Character: The output file name for the feature table.}
}
\value{
Writes the feature table to file.
}
\description{
The GenBank feature table is a 5 column table that is used to submit
annotations to GenBank (www.ncbi.nlm.nih.gov/WebSub/html/help/feature-table.html).
This function takes a BED file of annotations and converts it into the feature
table format for GenBank submission.
}
\details{
At this stage, the function's use is largely limited to simple annotations of
complete seqeunces and is restricted to genes, mRNA, tRNA, rRNA, and misc_features.
Coding of partial sequences and more complex annotations may be added in the future.
}
\examples{
library(genomalicious)

# Create a link to raw external datasets in genomalicious
genomaliciousExtData <- paste0(find.package('genomalicious'), '/extdata')

# We will create a feature table of gene and coding sequences for the protein-coding
# genes from the Bathygobius cocosensis mitogenome.
bcocoFtRefs <- fread(paste0(genomaliciousExtData, '/data_Bcocosensis_mito_features.csv'))
bcocoBedPath <- paste0(genomaliciousExtData, '/data_Bcocosensis_mito_annots.bed')

# Take a look
bcocoFtRefs
bcocoBedPath
readLines(bcocoBedPath)

# Run
bed2featuretab(bcocoBedPath, bcocoFtRefs, 'Bcoco_feature_table.txt')

}
