% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_poolfstat.R
\name{boot_poolfstat}
\alias{boot_poolfstat}
\title{Bootstrap FST values from \code{poolfstat}}
\usage{
boot_poolfstat(dat, pool.info, num.sims = 100)
}
\arguments{
\item{dat}{Data table: Contains read counts, e.g. like that been
produced by the function \code{vcf2DT}. Must contain all the following columns:
\code{$CHROM}, \code{$POS}, \code{$REF}, \code{$ALT}, \code{$LOCUS}, \code{$POOL},
\code{$DP}, \code{$RO}.}

\item{pool.info}{Data table: Contains the sample sample sizes (number of diploids) for
for each unique pool listed in \code{dat$POOL}. Requires two columns: \nr
\enumerate{
   \item \code{$POOL} The pools listed in \code{$dat$POOL}.
   \item \code{$INDS} The number of diploid individuals for the pools.
}}

\item{num.sims}{Integer: The number of bootstrap simulations to run. Default = 100.}
}
\value{
Returns a vector of FST values produced by bootstrapping loci in the
original read count dataset, \code{dat}.
}
\description{
Takes a data table of read counts and bootstraps the estimate of FST
calcualted from the function \code{poolfstat::computeFST}. Also requires
pool size information.
}
\examples{
#' # Load in the pool metadata and reads
data(pgposerInfo)
data(pgposerReads)

# Subset to keep only Rep1 reads.
X <- pgposerReads[grep(pattern='Rep1', x=pgposerReads$SAMPLE)]

# Need to add pool ID.
X$POOL <- unlist(lapply(strsplit(X$SAMPLE, '_'), function(X){ return(X[1]) }))

# Bootstrap FST
bootFST <- boot_poolfstat(pgposerReads, pgposerInfo, 100)

}
