% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_cutter.R
\name{circle_cutter}
\alias{circle_cutter}
\title{Cut a replicated circular sequence to generate a linear sequence}
\usage{
circle_cutter(query_seq, motif)
}
\arguments{
\item{query_seq}{Character: The query sequence.}

\item{motif}{Character, the sequence used to identify cut points in the
circular chromosome and to remove replicated regions. See Details.}
}
\value{
Returns the subset sequence as a character string.
}
\description{
This function cuts a circular sequence using a character string motif to
generate a single linear sequence. This is primarily intended for "read-through"
type sequences generated, for example, from assembly or long-read sequencing
of circular genomes. Such sequences typically have replicated regions
that need to be removed to generate a singular un-replicated linear sequence.
}
\details{
The argument \code{motif} is used to identify a starting and end
positions to cut the circular chromosome and remove replciated regions.
If for example \code{motif=='AATTGGCC'} and the sequence in question was: \cr
\cr
AATTGGCC ACTATCTGCTAGCTAGCATAGCATCGATCAGCATGACGCGCAA AATTGGCC \cr
\cr
The function will cut the sequence like so (marked with '|'): \cr
\cr
| AATTGGCC ACTATCTGCTAGCTAGCATAGCATCGATCAGCATGACGCGCAA | AATTGGCC
}
\examples{
x <- 'AATTGGCCACTATCTGCTAGCTAGCATAGCATCGATCAGCATGACGCGCAAAATTGGCC'

# Find character motif that is repeated
motif_hits <- circularity_test(x, word_size = 8)
motif_seq <- substr(x, motif_hits[1,1], motif_hits[1,2])

circle_cutter(query_seq = x, motif=motif_seq)

}
