% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da2DT.R
\name{da2DT}
\alias{da2DT}
\title{Convert a \code{lda} object to a long-format \code{data.table}}
\usage{
da2DT(daObj, sampVec, obsPops)
}
\arguments{
\item{daObj}{Lda: An object produced by \code{lda}.}

\item{sampVec}{Character: The sample IDs, which need to match the sample
order in \code{predict(daObj)$x}.}

\item{obsPops}{Character/Factor: The observed population IDs, which need to
match sample order in \code{predict(daObj)$x}.}
}
\value{
Returns a long-format data.table with columns \code{$POP}, the
population IDs, \code{$SAMPLE}, the sample IDs, \code{$AXIS}, the LD axis ID,
and \code{$SCORE}, the score on the LD axis.
}
\description{
Convert a \code{lda} object to a long-format \code{data.table}
}
\examples{
data(data_4pops)

PCA <- pca_genos(data_4pops, popCol='POP')
DA <- lda(PCA$pops ~ PCA$x[,1:3])

da2DT(DA, sampVec=rownames(PCA$x), obsPops=PCA$pops)

}
