% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadi_inputs_genos.R
\name{dadi_inputs_genos}
\alias{dadi_inputs_genos}
\title{Genertate dadi input from individual genotype data}
\usage{
dadi_inputs_genos(
  dat,
  sampCol = "SAMPLE",
  popCol = "POP",
  locusCol = "LOCUS",
  refCol = "REF",
  altCol = "ALT",
  genoCol = "GT",
  popSub = NULL,
  popLevels = NULL
)
}
\arguments{
\item{dat}{Data table: Contains population and genotype information. Genotypes
must be coded as '/' separated characters (e.g. '0/0', '0/1', or '1/1') or
integers of Alt allele counts (e.g. 0, 1, 2). Must contain the following columns:
\enumerate{
\item Sample ID (see argument \code{sampCol})
\item Population ID
\item Locus ID
\item Reference allele
\item Alternate alelle
\item Genotype
}}

\item{sampCol}{Character: Sample ID. Default = \code{'SAMPLE'}.}

\item{popCol}{Character: Population pool ID. Default = \code{'POP'}.}

\item{locusCol}{Character: Locus ID. Default = \code{'LOCUS'}.}

\item{refCol}{Character: Reference allele. Default = \code{'REF'}.}

\item{altCol}{Character: Alternate allele. Default = \code{'ALT'}.}

\item{genoCol}{Character: The genotype. Default = \code{'GT'}.}

\item{popSub}{Character: The populations to subset out of \code{popCol}. Default = \code{NULL}.}

\item{popLevels}{Character: An optional vector of the population pool IDs used
to manually specify the first and second population order. Default = \code{NULL}.}
}
\value{
Returns a data table in the dadi input format.
}
\description{
Creates an input file for the program dadi, described in Gutenkunst et al. (2009).
This function specifically uses individually sequenced genotypes to produce
inferred allele counts.
}
\examples{
library(genomalicious)

data(data_Genos)
datGt <- copy(data_Genos)

# Make the dadi input
dadi_inputs_genos(dat=datGt, popSub=c('Pop1', 'Pop2'))

}
\references{
Gutenkunst et al. (2009) Inferring the joint demographic history of multiply populations
from multidimensional SNP frequency data. PLoS Genetics: 10, e1000695.
}
