% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadi_inputs_pools.R
\name{dadi_inputs_pools}
\alias{dadi_inputs_pools}
\title{Genertate dadi input from pool-seq data}
\usage{
dadi_inputs_pools(dat, poolCol = "POOL", locusCol = "LOCUS",
  refCol = "REF", altCol = "ALT", freqCol = "PI", indsCol = "INDS",
  poolSub = NULL)
}
\arguments{
\item{dat}{Data table: Must contain columns with the following information,
\enumerate{
\item Population pool ID
\item Locus ID
\item Reference allele
\item Alternate alelle
\item Reference allele freuqency
\item Number of individuals per population pool
            }}

\item{poolCol}{Character: Population pool ID. Default = \code{'POOL'}}

\item{locusCol}{Character: Locus ID. Default = \code{'LOCUS'}}

\item{refCol}{Character: Reference allele. Default = \code{'REF'}}

\item{altCol}{Character: Alternate allele. Default = \code{'ALT'}}

\item{freqCol}{Character: The reference allele frequency. Default = \code{'FREQ'}.}

\item{indsCol}{Character: The number of individuals per population pool. Default = \code{'INDS'}.}

\item{poolSub}{Character: The pools to subset out of \code{poolCol}. Default = \code{NULL}.}
}
\value{
Returns a data table in the dadi input format. NOTE: The estimated number
of indivdiuals carrying an allele is rounded to nearest integer (e.g. 1.5 = 2, and 1.4 = 1),
with the exception when the number of individauls is < 1 but > 0, in which it is always rounded to 1.
}
\description{
Creates an input file for the program dadi, described in Gutenkunst et al. (2009).
}
\examples{
data(genomalicious_PoolPi)
genomalicious_PoolPi

dadi_inputs_pools(dat=genomalicious_PoolPi
                  , poolCol='POOL'
                  , locusCol='LOCUS'
                  , refCol='REF'
                  , altCol='ALT'
                  , freqCol='PI'
                  , indsCol='INDS'
                  , poolSub=c('Pop1', 'Pop2'))


}
\references{
Gutenkunst et al. (2009) Inferring the joint demographic history of multiply populations
from multidimensional SNP frequency data. PLoS Genetics: 10, e1000695.
}
