% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingBYsamples.R
\docType{data}
\name{dat}
\alias{dat}
\title{Plot missing data, by samples}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 208440 rows and 5 columns.}
\usage{
dat
}
\description{
Plot missing data, by samples
}
\keyword{datasets}
