% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadi_inputs_genos.R
\name{dat_input_genos}
\alias{dat_input_genos}
\title{Genertate dadi input from individual genotype data}
\usage{
dat_input_genos(dat, sampCol = "SAMPLE", popCol = "POP",
  locusCol = "LOCUS", refCol = "REF", altCol = "ALT", genoCol = "GT",
  popSub = NULL)
}
\arguments{
\item{dat}{Data table: Contains population and genotype information. Genotypes
must be coded as '/' separated characters (e.g. '0/0', '0/1', or '1/1') or
integers of Alt allele counts (e.g. 0, 1, 2). Must contain the following columns:
\enumerate{
\item Sample ID (see argument \code{sampCol})
\item Population ID
\item Locus ID
\item Reference allele
\item Alternate alelle
\item Genotyp
}}

\item{sampCol}{Character: Sample ID. Default = \code{'SAMPLE'}.}

\item{popCol}{Character: Population pool ID. Default = \code{'POP'}.}

\item{locusCol}{Character: Locus ID. Default = \code{'LOCUS'}.}

\item{refCol}{Character: Reference allele. Default = \code{'REF'}.}

\item{altCol}{Character: Alternate allele. Default = \code{'ALT'}.}

\item{genoCol}{Character: The genotype. Default = \code{'GT'}.}

\item{popSub}{Character: The populations to subset out of \code{poolCol}. Default = \code{NULL}.}
}
\value{
Returns a data table in the dadi input format.
}
\description{
Creates an input file for the program dadi, described in Gutenkunst et al. (2009).
}
\examples{
data(genomalicious_4pops)
datGt <- copy(genomalicious_4pops)

# Subset data for example (10 loci) and create artificial Ref and Alt
# nucleotides for the purpose of illustration
loci10 <- unique(datGt$LOCUS)[1:10]
datGt <- datGt[LOCUS \%in\% loci10]
ref10 <- sample(c('G', 'A', 'T', 'C'), 10, replace=TRUE)
names(ref10) <- loci10

alt10 <- sample(c('G', 'A', 'T', 'C'), 10, replace=TRUE)
names(alt10) <- loci10

datGt$REF <- ref10[match(datGt$LOCUS, names(ref10))]
datGt$ALT <- alt10[match(datGt$LOCUS, names(alt10))]

# Make the dadi input
dat_input_genos(dat=datGt, popSub=c('Pop1', 'Pop2'))

}
\references{
Gutenkunst et al. (2009) Inferring the joint demographic history of multiply populations
from multidimensional SNP frequency data. PLoS Genetics: 10, e1000695.
}
