% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_4pops.R
\docType{data}
\name{data_4pops}
\alias{data_4pops}
\title{An example of multi-population genotype data}
\format{
A data table with 144600 rows and 5 columns.
}
\usage{
data(data_4pops)
}
\description{
A dataset containing individual genotypes for 4 populations, each with 30
individuals, and 200 biallelic SNP loci. Simulated with FastSimCoal2 (Excoffier et al., 2013).
}
\details{
\itemize{
  \item \code{CHROM} = The contig ID.
  \item \code{POS} = The SNP position.
  \item \code{LOCUS} = The unique locus ID.
  \item \code{POP} = The population ID.
  \item \code{SAMPLE} = The sample ID: Ind[Pop]_[Number].
  \item \code{GT} = The genotype ID, coded as counts of the alternate allele.
}
}
\references{
Excoffier et al. (2013) Robust demographic inference from genomic SNP data.
PLoS Genetics: 9(10), e1003905.
}
\keyword{datasets}
