% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomalicious_Freqs.R
\docType{data}
\name{data_Freqs}
\alias{data_Freqs}
\title{An example matrix of allele frequencies}
\format{
A data table with 4 rows and 8 columns.
}
\usage{
data(data_Freqs)
}
\description{
A dataset containing allele frequency information for 4 populations
and 8 biallelic SNP loci.
}
\details{
Population ID are in \code{rownames(data_Freqs)}, and locus
ID are in \code{colnames(data_Freqs)}.
}
\keyword{datasets}
