% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PatchyGTs.R
\docType{data}
\name{data_PatchyGTs}
\alias{data_PatchyGTs}
\title{An example of a patchy genotype matrix}
\format{
A matrix with 100 rows and 100 columns.
}
\usage{
data(data_PatchyMat)
}
\description{
A matrix of genotypes with missing data.
}
\details{
Samples IDs are in \code{rownames(data_PatchyGTs)}, and locus
IDs are in \code{colnames(data_PatchyGTs)}. Missing data are
coded by \code{NA.}
}
\keyword{datasets}
