% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PoolInfo.R
\docType{data}
\name{data_PoolInfo}
\alias{data_PoolInfo}
\title{An example of pool-seq sample info}
\format{
A data table with 2 rows and 2 columns.
}
\usage{
data(data_PoolInfo)
}
\description{
A dataset containing the population pool ID and the number of individuals
used in a pool-seq experiment. \cr
\cr
}
\details{
\itemize{
  \item \code{POOL} = The population pool ID.
  \item \code{INDS} = The number of diploid individuals that contributed toward
    the population pool.
}
}
\references{
Gautier et al. (2013) Estimation of population allele frequencies from
next-generation sequencing data: pool-versus individual-based genotyping.
Molecular Ecology: 22(14) 3766–3779.
}
\keyword{datasets}
