% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decircularise_longread.R
\name{decircularise_longread}
\alias{decircularise_longread}
\title{Decircularise a circular chromosome generated from long reads}
\usage{
decircularise_longread(fastaIn, fastaOut, motif)
}
\arguments{
\item{fastaIn}{Character, the path to a FASTA file. Must only contain
one sequence.}

\item{fastaOut}{Character, the path to write out the cut FASTA.}

\item{motif}{Character, the sequence used to identify cut points in the
circular chromosome and to remove replicated regions. See Details.}
}
\description{
Long-read technologies can provide complete read throughs of circular
chromosomes. However, when the number of bases sequenced is greater than the
chromosome size, this can lead to sequence replication within the long reads.
This function makes the circule readthrough linear. It requires a
user-specified motif to identify replicated regions and remove them.
}
\details{
The argument \code{motif} is used to identify a starting and end
positions to cut the circular chromosome and remove replciated regions.
If for example \code{motif=='AATTGGCC'} and the sequence in question was: \cr
\cr
GCGCAA AATTGGCC ACTATCTGCTAGCTAGCATAGCATCGATCAGCATGAC GCGCAA AATTGGCC \cr
\cr
The function will cut the sequence like so (marked with '|'): \cr
\cr
GCGCAA | AATTGGCC ACTATCTGCTAGCTAGCATAGCATCGATCAGCATGAC GCGCAA | AATTGGCC
}
