% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT2fasta.R
\name{DT2fasta}
\alias{DT2fasta}
\title{Data table to FASTA file}
\usage{
DT2fasta(dat, fastaFile)
}
\arguments{
\item{dat}{Data table: Contains columns \code{$LOCUS} of locus names
and \code{$SEQ} of nucleotide sequences.}

\item{fastaFile}{Character: The path to the output FASTA file.}
}
\value{
Writes FASTA file.
}
\description{
Converts a data table of sequences (e.g. like that produced by
\code{fasta2DT()}) and writes a FASTA file
}
