% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT2poolfstat.R
\name{DT2poolfstat}
\alias{DT2poolfstat}
\title{Calculate FST with \code{poolfstat} from a data table of read counts}
\usage{
DT2poolfstat(dat, pool.info)
}
\arguments{
\item{dat}{Data table: Contains read counts, e.g. like that been
produced by the function \code{vcf2DT}. Must contain all the following columns:
\code{$CHROM}, \code{$POS}, \code{$REF}, \code{$ALT}, \code{$LOCUS}, \code{$POOL},
\code{$DP}, \code{$RO}.}

\item{pool.info}{Data table: Contains the sample sample sizes (number of diploids) for
for each unique pool listed in \code{dat$POOL}. Requires two columns: \nr
\enumerate{
   \item \code{$POOL} The pools listed in \code{$dat$POOL}.
   \item \code{$INDS} The number of diploid individuals for the pools.
}}
}
\value{
Returns a list with two indices: \code{$Fst} is the calculated FST among the
pools, and \code{$pooldat} is the \code{poolfstat} object used to generate said FST.
}
\description{
Takes a data table of read counts and creates an object of class
\code{poolfstat}. The FST for the pools in the data table is calculated using
the function \code{poolfstat::computeFST}. Also requires pool size information.
}
\examples{
# Load in the pool metadata and reads
data(pgposerInfo)
data(pgposerReads)

# Subset to keep only Rep1 reads.
X <- pgposerReads[grep(pattern='Rep1', x=pgposerReads$SAMPLE)]

# Need to add pool ID.
X$POOL <- unlist(lapply(strsplit(X$SAMPLE, '_'), function(X){ return(X[1]) }))

# Calculate FST using poolfstat
DT2poolfstat(X, pgposerInfo)

}
