% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT2RADdata.R
\name{DT2RADdata}
\alias{DT2RADdata}
\title{Convert a long-format data table into polyRAD's RADdata object class.}
\usage{
DT2RADdata(
  dat,
  chromCol = "CHROM",
  posCol = "POS",
  locusCol = "LOCUS",
  sampCol = "SAMPLE",
  refCol = "REF",
  roCol = "RO",
  altCol = "ALT",
  aoCol = "AO",
  possPloidy = list(2L),
  sampPloidy = 2L,
  contamRate = 0.001
)
}
\arguments{
\item{dat}{Data.table: A data table of read counts for genotypes loci.
Expects the columns:
\enumerate{
   \item The chromosome (contig) ID (see param \code{chromCol}).
   \item The position information (see param \code{posCol}).
   \item The locus ID (see param \code{locusCol}).
   \item The sample ID (see param \code{sampCol}).
   \item The reference allele nucleotides (see param \code{refCol}).
   \item The reference allele read counts (see param \code{roCol}).
   \item The alternate allele nucleotides (see param \code{altCol}).
   \item The alternate allele read counts (see param \code{aoCol}).
   (see param \code{freqCol})
}}

\item{chromCol}{Character: The column name with the chromosome information.
Default = \code{'CHROM'}.}

\item{posCol}{Character: The column name with the position information.
Default = \code{'POS'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}. Only needed when \code{type=='genos'}.}

\item{refCol}{Character: The column with the reference allele nucleotide information.
Default = \code{'REF'}.}

\item{roCol}{Character: The column with the reference allele read count information.
Default = \code{'RO'}.}

\item{altCol}{Character: The column with the alternate allele nucleotide information.
Default = \code{'ALT'}.}

\item{aoCol}{Character: The column with the alternate allele read count information.
Default = \code{'AO'}.}

\item{possPloidy}{List: A list of integers or numerics that represent the unique
ploidy values in the dataset. Default is \code{list(2)}, i.e., all samples have
a ploidy of 2. A list of \code{list(2, 4)} would represent possible ploidies of
2 and 4.}

\item{sampPloidy}{Integer/Numeric: Either a single value or a named vector of values
for each sample. Default is a single value, 2, i.e., all samples have a ploidy
of 2. A vector \code{c('Ind1'=2, 'Ind2'=4, 'Ind3'=2)}, for example, is a vector
of ploidies for three individuals, with ploidy values of 2, 4, and 2 for individuals 1, 2
and 3, respectively.}

\item{contamRate}{Numeric: The contamination rate. Default is 0.001.}
}
\description{
Used to generate the RADdata class object from the \code{polyRAD} package
described in Clark et al. (2019).
}
\examples{
library(genomalicious)
data(data_Genos)

# Using a single ploidy
RD1 <- DT2RADdata(data_Genos, sampPloidy=2)

# Using a vector of ploidies.
samps_uniq <- unique(data_Genos$SAMPLE)
samps_ploid <- rep(2, length(samps_uniq))
names(samps_ploid) <- samps_uniq

samps_ploid

RD2 <- DT2RADdata(data_Genos, sampPloidy=samps_ploid)

Specifying multiple ploidies.
samps_ploid[20:40] <- 4

samps_ploid \%>\% table

RD3 <- DT2RADdata(data_Genos, possPloidy=list(2, 4), sampPloidy=samps_ploid)

}
\references{
Clark et al. (2019). G3. DOI: 10.1534/g3.118.200913
}
