% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_sim_Amatrix.R
\name{family_sim_Amatrix}
\alias{family_sim_Amatrix}
\title{The additive genetic relatedness matrix for simulated individuals}
\usage{
family_sim_Amatrix(simFam, type)
}
\arguments{
\item{simFam}{Character: A vector of simulated indivdiual sample IDs}
}
\value{
Returns a symmetrical matrix of additive genetic relationships
between all pairs of individuals.
}
\description{
Takes individuals generated from \code{family_sim_data} and returns an
additive genetic relatedness matrix
}
\examples{
library(genomalicious)
data(data_Genos)

# Subset Pop1 genotypes
genosPop1 <- data_Genos[POP=='Pop1', c('SAMPLE', 'LOCUS', 'GT')]

# Get the allele frequencies for Pop1
freqsPop1 <- genosPop1[, .(FREQ=sum(GT)/(length(GT)*2)), by=LOCUS]

# Simulate 100 families
simFamily <- family_sim_data(
   freqData=freqsPop1,
   locusCol='LOCUS',
   freqCol='FREQ',
   numSims=100
)

### THE SIMULATED ADDITIVE GENETIC RELATEDNESS MATRIX
simARM <- family_sim_Amatrix(unique(simFamily$SAMPLE))

simARM[1:8, 1:8]

}
