% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_sim_annotate.R
\name{family_sim_annotate}
\alias{family_sim_annotate}
\title{Annotate pairwise matrix of simulated family members}
\usage{
family_sim_annotate(simFam, type)
}
\arguments{
\item{simFam}{Matrix or Character: Either a pairwise matrix or a character vector of
individual sample IDs. If a pairwise matrix, the matrix must be symmetrical with sample
IDs in row and column names: the diagonal compares individuals to themselves,
and off-diagonal is comparisons between individuals; must set \code{type=="matrix"}.
If a character vector of sample IDs, must set \code{type=="character"}.}

\item{type}{Character: Either "matrix" or "id_vec".}
}
\value{
Returns a long-format data.table with the columns:
\enumerate{
   \item \code{$SIM}, the simulation number for pairs of simulated individuals,
   or 'NA' for the pairs of observed individuals.
   \item \code{$SAMPLE1}, the sample ID for the first individual.
   \item \code{$SAMPLE2}, the sample ID for the second individual.
   \item \code{$FAMILY}, the familial relationship for simulated individuals.
   \item \code{$Y}, the value in the cell contents from the input matrix, if
   \code{type=='matrix'}.
}
}
\description{
Takes individuals generated from \code{family_sim_data} and annotates
familial relationships across all pairs of individuals. Takes either a pairwise
matrix of individuals or a character vector.
}
\examples{
library(genomalicious)
data(data_Genos)

# Subset Pop1 genotypes
genosPop1 <- data_Genos[POP=='Pop1', c('SAMPLE', 'LOCUS', 'GT')]

# Get the allele frequencies for Pop1
freqsPop1 <- genosPop1[, .(FREQ=sum(GT)/(length(GT)*2)), by=LOCUS]

# Simulate 100 families
simFamily <- family_sim_data(
   freqData=freqsPop1,
   locusCol='LOCUS',
   freqCol='FREQ',
   numSims=100
)

### THE SIMULATED GENOMIC RELATIONSHIPS MATRIX
library(AGHmatrix)

# Convert simulated families into a genotype matrix
simGenosMat <- DT2Mat_genos(simFamily)

# Calculate the GRM
simGRM <- Gmatrix(simGenosMat, method='Yang', ploidy=2)

### THE FAMILAL ANNOTATIONS
# From a pairwise matrix (the GRM)
family_sim_annotate(simFam = simGRM, type = 'matrix')

# From a vector of unique individual IDs
family_sim_annotate(simFam = rownames(simGRM), type = 'id_vec')

}
