% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2DT.R
\name{fasta2DT}
\alias{fasta2DT}
\title{FASTA file to data table}
\usage{
fasta2DT(fastaFile)
}
\arguments{
\item{fastaFile}{Character: The path to the input FASTA file.}
}
\value{
A data table with the columns \code{$LOCUS} (the locus name) and \code{$SEQ} (the sequence).
}
\description{
Reads a FASTA file and converts to a data table.
}
