% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsimcoal2_snps2dips.R
\name{fastsimcoal2_snps2dips}
\alias{fastsimcoal2_snps2dips}
\title{Convert \code{fastsimcoal2} SNPs into diploid genotypes}
\usage{
fastsimcoal2_snps2dips(arpFile, padIDs = TRUE)
}
\arguments{
\item{arpFile}{Character: The name of the \code{.arp} file produced from
the \code{fastsimecoal2} simulations.}

\item{padIDs}{Logical: Should ID numbers be padded with '0'? Useful for
sorting character values. Defalt is TRUE.}
}
\value{
Returns a data table with the following columns:
\enumerate{
   \item \code{$POP} Population ID (simualted samples, "Sample" 1 -> n;
   takes form \code{pop_[ID]}).
   \item \code{$SAMPLE} Individual ID (the diploid individual from each
   population; takes form \code{ind_[pop]_[ID]}).
   \item \code{$LOCUS} The locus ID (takes form \code{locus_[chrom]_[ID]}).
   \item \code{$CHROM} The chromosome ID (takes form \code{chrom_[ID]})
   \item \code{$GT} The genotype, scored as '/' separated alleles, where 0
   is ancestral, and 1 is derived.
}
}
\description{
Takes in a \code{.arp} file produced from a \code{fastsimcoal2} (Excoffier et al. 2013)
simulation of SNP data and produces diploid genotypes from the sampled haploid genomes.
NOTE: This function was developed for even sample sizes (i.e. total number
of haploid genomes per population is divisible by 2) and for simulations where
all loci are polymorphic. Users should proceed with caution if their simulated data
does not fit these specifications.
}
\details{
Diploid genotypes are created by combining pairs of haploid genomes
from each population (i.e. "Sample" 1 -> n in \code{fastsimcoal2}).
}
\references{
Excoffier et al. (2013) Robust demographic inference from genomic and SNP data. PLoS Genetics.
}
