% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_depth.R
\name{filter_depth}
\alias{filter_depth}
\title{Filter loci by sequencing depth}
\usage{
filter_depth(dat, minDP = 10, maxDP = NA)
}
\arguments{
\item{dat}{Data table: The sequencing read information, must contain the columns:
\code{$LOCUS} = the locus ID; and \code{$DP} = the depth of sequencing reads.}

\item{minDP}{Integer: The minimum sequencing depth. Loci below this value are flagged
as 'bad' loci. Default = \code{10}. Must be specified; set \code{minDP = 0} if there is no
minimum number of reads.}

\item{maxDP}{Integer: The maximum sequencing depth. Loci above this value are flagged
as 'bad' loci. Default = \code{NA}.}
}
\value{
Returns a character vector of locus names in \code{dat$LOCUS} that conform
to the read depth values specified, i.e. 'good' loci.
}
\description{
Parses a data table of sequencing read information to determine which loci
conform to desired sequencing depth values.
}
\examples{
data(pgposerReads)

# Look at the sequencing coverage
pgposerReads

# Exclude loci with coverage < 30 reads
min30 <- filter_depth(pgposerReads, minDP=30)
min30
pgposerReads[LOCUS \%in\% min30]

# Exclude loci with coverage < 30 and > 200 reads
min30max200 <- filter_depth(pgposerReads, minDP=30, maxDP=200)
min30max200
pgposerReads[LOCUS \%in\% min30max200]

# Alternatively, subset data.table to only contain the
# bad loci with coverage < 30 reads
pgposerReads[!(LOCUS \%in\% min30)]

}
