% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_missing_loci.R
\name{filter_missing_loci}
\alias{filter_missing_loci}
\title{Filter missing data by loci}
\usage{
filter_missing_loci(
  dat,
  missing,
  type = "genos",
  method = "samples",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  popCol = "POP",
  genoCol = "GT",
  freqCol = "FREQ"
)
}
\arguments{
\item{dat}{Data table: The must contain the columns:
\enumerate{
   \item The sample ID (see param \code{sampCol}), for genotype datasets only.
   \item The locus ID (see param \code{locusCol}).
   \item The population ID (see param \code{popCol}).
   \item The genotypes (see param \code{genoCol}), or the allele frequencies
   (see param \code{freqCol})
}}

\item{missing}{Numeric: The proportion of missing data a locus, a value between
0 and 1.}

\item{type}{Character: Is \code{dat} a data table of genotypes (\code{'genos'})
or a data table of allele frequencies (\code{'freqs'})? Default = \code{'genos'}.}

\item{method}{Character: The method by which missingness filtering is performed.
Only valid when filtering is performed on genotypes (\code{type=='genos'}).
One of \code{'samples'}, or \code{'pops'}. Default = \code{'samples'}.
For \code{'samples'}, missingness is calculated across all sampled individuals
(irrespective of their populations) for genotypes at each locus. If the
missingness summed across samples is greater than the threshold, the locus
will be discarded. For \code{'pops'}, missingness at a locus is calculated
per population. If any populations has missingness above the threshold at
a locus, then that locus will be removed.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}. Only needed when \code{type=='genos'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{popCol}{Character: The column name with population information.
Default = \code{'POP'}.}

\item{genoCol}{Character: The column name with the genotype information.
Missing genotypes are encoded with an \code{NA}. Default = \code{'GT'}.
Only needed when \code{type=='genos'}.}

\item{freqCol}{Character: The column name with the allele frequency information.
Missing frequencies are encoded with an \code{NA}. Default = \code{'freqCol'}.
Only needed when \code{type=='freqs'}.}
}
\value{
Returns a character vector of locus names in \code{dat$LOCUS} that conform
to the missingness threshold (<= to the value of \code{missing}).
}
\description{
Parses a data table of genotypes/allele frequencies and returns a list of
loci that conform to a desired missing data threshold.
}
\details{
Note, it is assumed that missing data values have already been put is as
an \code{NA}. If this is not done in advance, this function will not produce
the expected results.

If \code{type=='genos'}, then your output will depend on how you
specify the \code{method} argument. If \code{type=='freqs'}, then there is
just one output, those loci with missing data less than the \code{missing} threshold.
}
\examples{
library(genomalicious)

simMiss <- data_Genos \%>\% copy()
simMiss$GT[sample(1:nrow(simMiss), 0.1*nrow(simMiss), replace=FALSE)] <- NA

filter_missing_loci(simMiss, 0.10)

}
