% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_missing_units.R
\name{filter_missing_units}
\alias{filter_missing_units}
\title{Filter missing data by units (sampled individuals or populations)}
\usage{
filter_missing_units(
  dat,
  missing,
  type = "genos",
  method = "samples",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  popCol = "POP",
  genoCol = "GT",
  freqCol = "FREQ"
)
}
\arguments{
\item{dat}{Data table: The must contain the columns:
\enumerate{
   \item The individual sample ID (see param \code{sampCol}), for genotype
   datasets only.
   \item The locus ID (see param \code{locusCol}).
   \item The population ID (see param \code{popCol}).
   \item The genotypes (see param \code{genoCol}), or the allele frequencies
   (see param \code{freqCol})
}}

\item{missing}{Numeric: The proportion of missing data a locus, a value between
0 and 1.}

\item{type}{Character: Is \code{dat} a data table of genotypes (\code{'genos'})
or a data table of allele frequencies (\code{'freqs'})? Default = \code{'genos'}.}

\item{method}{Character: The method by which missingness filtering is performed.
Only valid when filtering is performed on genotypes (\code{type=='genos'}).
One of \code{'samples'}, or \code{'pops'}. Default = \code{'samples'}.
For \code{'samples'}, missingness is calculated across all loci for each
sampled individual. Any individual with missingness greater than the threshold
will be discarded. For \code{'pops'}, a mean missingness is calculated for
each populations, across all sampled individuals within those populations and
across all loci. Populations with missingness greater than the threshold will
be discarded.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}. Only needed when \code{type=='genos'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{popCol}{Character: The column name with population information.
Default = \code{'POP'}.}

\item{genoCol}{Character: The column name with the genotype information.
Missing genotypes are encoded with an \code{NA}. Default = \code{'GT'}.
Only needed when \code{type=='genos'}.}

\item{freqCol}{Character: The column name with the allele frequency information.
Missing frequencies are encoded with an \code{NA}. Default = \code{'freqCol'}.
Only needed when \code{type=='freqs'}.}
}
\value{
Returns a character vector of samples names in \code{dat[[sampCol]]} or
populations in \code{dat[[popCol]]} that conforms to missingness threshold
(<= to the value of \code{missing}).
}
\description{
Parses a data table of genotypes/allele frequencies and returns a list of
units (sampled individuals or populations) that conform to a desired missing data threshold.
}
\details{
Note, it is assumed that missing data values have already been put is as
an \code{NA}. If this is not done in advance, this function will not produce
the expected results.

If \code{type=='genos'}, then your output will depend on how you
specify the \code{method} argument. If \code{type=='freqs'}, then there is
just one output, those populations with missing data less than the \code{missing} threshold.
}
\examples{
library(genomalicious)

simMiss <- data_4pops \%>\% copy()
simMiss$GT[sample(1:nrow(simMiss), 0.1*nrow(simMiss), replace=FALSE)] <- NA

filter_missing_units(simMiss, missing=0.10)
filter_missing_units(simMiss, missing=0.10, method='pops')

}
