% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_unlink.R
\name{filter_unlink}
\alias{filter_unlink}
\title{Filter for unlinked loci}
\usage{
filter_unlink(dat)
}
\arguments{
\item{dat}{Data table: The sequencing read information, must contain the columns:
\code{CHROM} = the chromosome ID, i.e. contigs; and \code{LOCUS} = the locus ID.}
}
\value{
Returns a character vector of locus names in \code{dat$LOCUS} that are not on the same
contig in \code{dat$CHROM}.
}
\description{
Parses a data table of sequence read information to randomly draw one locus
per contig.
}
