% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs_DT2Mat.R
\name{freqs_DT2Mat}
\alias{freqs_DT2Mat}
\title{Convert a data table of allele frequencies into a matrix (or vice versa)}
\usage{
freqs_DT2Mat(dat, popCol = "POOL", lociCol = "LOCUS",
  freqCol = "FREQ", flip = FALSE)
}
\arguments{
\item{dat}{Data table or Matrix: The object to transform. If this is a long data table
# of allele frequencies, then three columns are required:
\enumerate{
   \item (1) The population ID (see param \code{popCol}).
   \item (2) The locus ID (see param \code{lociCol}).
   \item (3) The Ref allele frequency (see param \code{freqCol}).
}
The population pool column serves as the pivot point to convert the long data table into a wide matrix.
If convertin from a frequency matrix to a data table, see argument \code{flip}.}

\item{popCol}{Character: The column name with the population information.
Default = \code{'POOL'} (e.g. for pool-seq data).}

\item{lociCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{freqCol}{Character: The column name with the Ref allele frequency.
Default = \code{'FREQ'}.}

\item{flip}{Logical: Instead of converting a (long) data table to a (wide) matrix,
should a (wide) matrix be converted into a (long) data table? Default = \code{FALSE}.
If \code{TRUE}, then param \code{dat} must be a matrix, with loci names as column headers,
population IDs in the row names, and frequencies in the cells. When \code{TRUE}, params
\code{popCol}, \code{lociCol}, and \code{freqCol} become void.}
}
\value{
When \code{flip=FALSE}, converts a data table into a frequency matrix. When
\code{flip=TRUE}, converts a matrix into a data table with three columns: (1) \code{$POOL},
the population ID (as for pool-seq data); (2) \code{$LOCUS}, the locus ID; and (3) \code{FREQ},
the Ref allele frequency.
}
\description{
Takes a data table of allele frequencies in long format and converts it into
a matrix in wide format (loci in columns and populations in rows). The reverse
can also be done.
}
\examples{
data(pgposerPi)

# Convert a long data table to a wide matrix
freqMat <- freq_DT2Mat(pgposerPi, popCol='POOL', lociCol='LOCUS', freqCol='PI', flip=FALSE)

# Convert a wide matrix back to a data table
freqDT <- freq_DT2Mat(freqMat, flip=TRUE)

export
}
