% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_calc.R
\name{fst_calc}
\alias{fst_calc}
\title{Calculate Weir and Cockerham's FST from genotypes or allele frequencies}
\usage{
fst_calc(
  dat,
  type,
  popCol = "POP",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT",
  freqCol = "FREQ",
  indsCol = "INDS",
  global = TRUE,
  pairwise = FALSE,
  permute = FALSE,
  numIters = 100
)
}
\arguments{
\item{dat}{Data table: A long format data table of biallelic genotypes,
coded as '/' separated alleles ('0/0', '0/1', '1/1'), or counts
of the Alt alleles (0, 1, 2, respectively).
Columns required for \strong{both} genotypes and allele frequencies:
\enumerate{
   \item The population ID (see param \code{popCol}).
   \item The locus ID (see param \code{locusCol}).
}
Columns required only for \strong{genotypes}:
\enumerate{
   \item The sample ID (see param \code{sampCol}).
   \item The genotypes (see param \code{genoCol}).
}
Columns required only for \strong{allele frequencies}:
\enumerate{
   \item The allele frequencies (see param \code{freqCol}).
   \item The number of individuals used to obtain the allele frequency
   estimate (see param \code{indsCol}).
}}

\item{type}{Character: One of \code{'genos'} or \code{'freqs'}, to calculate
FST on genotype or allele frequency data, respectively.}

\item{popCol}{Character: The column name with the population information.
Default is \code{'POP'}.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default is \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default is \code{'GT'}.}

\item{freqCol}{Character: The column name with the allele freuqency information.
Default is \code{'FREQ'}.}

\item{indsCol}{Character: The column name with the number of individuals
contributing to the allele freuqency estimate. Default is \code{indsCol}.}

\item{global}{Logical: Calculate the global FST across all populations?
Default is \code{TRUE}. Note, \code{global==TRUE} and \code{pairwise==TRUE}
is not permitted.}

\item{pairwise}{Logical: Calculate pairwise FST between populations?
Default is \code{FALSE}. Note, \code{global==TRUE} and \code{pairwise==TRUE}
is not permitted.}

\item{permute}{Logical: Should permutations be performed to test the
statistical significance of FST? Default is \code{FALSE}. Can only be performed
on genotype data, i.e., \code{type=='genos'}.}

\item{numIters}{Integer: The number of permutations to perform. Default is 100.}
}
\value{
A list is returned with three indexes.

The first index is \code{$locus}, the locus-specific FST. This is a data table
object with the columns \code{$LOCUS}, the locus ID, \code{$NUMER}, the
numerator variance component, \code{$DENOM}, the denominator variance component,
and \code{$FST}, the locus specific FST. Additionally, if pairwise population
estimates have been requested, \code{pairwise==TRUE}, then there are
\code{$POP1} and \code{$POP2}, which represent the two populations tested.

The second index is \code{$genome}, the genome-wide FST. If the global estimate
has been requested, \code{global==TRUE}, then this is just a single value;
the estimate across all populations. If pairwise esimates were requested,
\code{pairwise==TRUE}, then there are \code{$POP1} and \code{$POP2}, which
represent two populations tested.

The third index is \code{$permute}, the permutation results. This index will
be \code{NULL} when frequencies are used, i.e., \code{type=='freqs'}, and
will only contain data if \code{type=='genos'} and \code{permute==TRUE}.
\code{$permute} is itself a list, with two subindexes:
\enumerate{
   \item \code{$fst}: The permuted FST values. If \code{global==TRUE}, then
   this will simply be a numeric vector of permuted global FST values.
   If \code{pairwise==TRUE}, then this will be a data table with columns
   \code{$POP1}, \code{$POP2}, and \code{$FST}.

   \item \code{$pval}: The permuted p-values. If \code{global==TRUE}, then
   this will single p-value for the global FST. If \code{pairwise==TRUE},
   then this will be a data table with columns \code{$POP1}, \code{$POP2},
   and \code{$PVAL}.
}
}
\description{
Function takes a genotypes or allele frequencies in a long-format data table
and calculates Weir & Cockerhams FST, i.e. theta (Weir & Cockerham, 1984).
Permutations can be used to test statistical significance of FST in
genotype data sets.
}
\details{
Permutation tests involve random shuffling of individuals
among populations, recalculating FST, and testing the hypothesis
that the permuted FST > observed FST. The p-value represents the
proportion of iterations that were "TRUE" to this expression. That is,
if no permuted values are greater than the observed, p=0. Likewise, if all
the permuted values are greater than the observed, p=1.
}
\examples{
library(genomalicious)

data(data_Genos)
data(data_PoolFreqs)

### Genotypes, global, with 50 permutations
genoGlobal <- fst_calc(
   data_Genos,
   type='genos',
   global=TRUE,
   permute=TRUE,
   numIters=50)

# Locus specific estimates
genoGlobal$locus

# Genome-wide estimate
genoGlobal$genome

# Permutations
genoGlobal$permute$fst
genoGlobal$permute$pval

### Genotypes, pairwise, no permutations
genoPairs <- fst_calc(
   data_Genos,
   type='genos',
   global=FALSE,
   pairwise=TRUE
   )

# Locus specific estimates
genoPairs$locus

# Genome-wide estimates
genoPairs$genome

# Permutations (there are none!)
genoPairs$permute

### Genotypes, pairwise, with 10 permutations
genoPairsPerms <- fst_calc(
   data_Genos,
   type='genos',
   global=FALSE,
   pairwise=TRUE,
   permute=TRUE,
   numIters=10
   )

# Permuted FST
genoPairsPerms$permute$fst

# Permuted p-value
genoPairsPerms$permute$pval

### Frequencies, pairwise, with 10 permutations
# Note columns in data_PoolFreqs
colnames(data_PoolFreqs)

# There is no $POP column, but there is a $POOL column that contains
# the population pool ID. Need to manually specify this.
freqPairs <- fst_calc(
   data_PoolFreqs,
   type='freqs',
   global=FALSE,
   pairwise=TRUE,
   popCol='POOL'
)

}
