% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_calc.R
\name{fstat_permute}
\alias{fstat_permute}
\title{Permute the genome-wide F-statistics for genotype data}
\usage{
fstat_permute(dat, fstat, numPerms, numCores = 1)
}
\arguments{
\item{dat}{Data table: In long-format, requires columns \code{$POP},
\code{$SAMPLE}, \code{$LOCUS} and \code{$GT} for genotypes;
requires \code{$POP}, \code{$LOCUS}, \code{$FREQ}, and \code{$INDS} for
allele frequencies}

\item{fstat}{Character: A vector containing the F-statistics to calculate
when the data are genotypes. Must include one of \code{"FST"},
\code{"FIS"}, or \code{"FIT"}.}

\item{numPerms}{Integer: The number of permutations to perform.}

\item{numCores}{Integer: The number of cores to use. Default = 1. When 1,
this will be a single core processes using lapply. If >1, then this will
utilise the doParallel package to run on a mini cluster.}
}
\value{
Returns a data table of permuted F-statistics for each permutation.
}
\description{
Uses permutation of individuals among populations to generate a null
distribution of F-statistics. Assumes biallelic SNP genotypes. \cr\cr
Note, this function is not exported and is an internal function for
\code{fstat_calc}.
}
\examples{
data(data_Genos)

fstat_permute(data_Genos, 'FST', 100, 1)

fstat_permute(data_Genos, c('FST','FIT','FIS'), 100, 3)
}
\keyword{internal}
