% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_calc.R
\name{fstat_subfun}
\alias{fstat_subfun}
\title{Subfunction for calculating F-statistics}
\usage{
fstat_subfun(dat, type, fstat = NULL)
}
\arguments{
\item{dat}{Data table: In long-format, requires columns \code{$POP},
\code{$SAMPLE}, \code{$LOCUS} and \code{$GT} for genotypes;
requires \code{$POP}, \code{$LOCUS}, \code{$FREQ}, and \code{$INDS} for
allele frequencies}

\item{type}{Character: either \code{"genos"} or \code{"freqs"}.}

\item{fstat}{Character: a vector containing the F-statistics to calculate
when the data are genotypes. Must include one of \code{"FST"},
\code{"FIS"}, or \code{"FIT"}.}
}
\value{
Returns a list with \code{$genome}, the genome-wide F-statistic, and
\code{$locus}, the per locus F-statistic.
}
\description{
Calculates F-statistics from genotype or allele frequency data in a
long-format data table as per Weir & Cockerham (1984). Assumes that all
values are for a single biallelic SNP locus. \cr\cr Note, this function is
not exported and is an internal function for \code{fstat_calc}.
}
\examples{
data_Genos \%>\%
   fstat_subfun(., type='genos', fstat=c('FST','FIS','FIT'))

left_join(data_PoolFreqs, data_PoolInfo) \%>\%
   setnames(., 'POOL', 'POP') \%>\%
   fstat_subfun(., type='freqs')
}
\keyword{internal}
