% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC.R
\name{fstWC}
\alias{fstWC}
\title{Calculate Weir and Cockerham's FST}
\usage{
fstWC(freqMat, sampMat, pairs = FALSE)
}
\arguments{
\item{freqMat}{Matrix: Ref allele counts. Rows = populations,
columns = loci; make sure both are named. Row names used to label output FST matrix.}

\item{sampMat}{Matrix: Number of sampled individuals. Rows = populations,
columns = loci.}

\item{pairs}{Logical: Should pairwise FSTwc be calculated (TRUE) or the
among population FSTwc (FALSE)?}
}
\value{
A list with two indices. \code{$fst.mean} contains the mean FST. In pairwise analyses,
this is a data table with population pairs. \code{$fst.locus} contains the locus-specific
FST values. If analysis was not pairwise, this is a vector, but for pairwise analyses,
this is a data table with population pairs.
}
\description{
Calculate Weir and Cockerham's FST
}
\examples{
freqMat <- genomaliciousFreqs
sampMat <- matrix(rep(30, 32), nrow=4, ncol=8)
rownames(sampMat) <- paste0('Pop', 1:4)
colnames(sampMat) <- colnames(freqMat); rownames(sampMat) <- rownames(freqMat)

fstWC(freqMat, sampMat, pairs=FALSE)
fstWC(freqMat, sampMat, pairs=TRUE)

}
\references{
Weir, Cockerham (1984) Estimating F-statistics for the analysis of population structure. Evol. \n
Weir, Hill (2002) Estimating F-statistics. Annu. Rev. Genet
}
