% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_boot.R
\name{fstWC_boot}
\alias{fstWC_boot}
\title{Bootstrap Weir & Cockerham's FST}
\usage{
fstWC_boot(freqMat, sampMat, boots = 100)
}
\arguments{
\item{freqMat}{Matrix: Ref allele counts. Rows = populations,
columns = loci; make sure both are named. Row names used to label output FST matrix.}

\item{sampMat}{Matrix: Number of sampled individuals. Rows = populations,
columns = loci.}

\item{boots}{Integer: The number of bootstrap replicates. Default = 100.}
}
\value{
Returns a vector bootstrapped mean FST values (estimated across
bootstrapped resampled loci).
}
\description{
Use a bootstrap resampling procedure to estimate the distribution
of values around observed FST value.
}
\references{
Weir, Cockerham (1984) Estimating F-statistics for the analysis of population structure. Evol. \cr
Weir, Hill (2002) Estimating F-statistics. Annu. Rev. Genet.
}
