% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_freqs.R
\name{fstWC_freqs}
\alias{fstWC_freqs}
\title{Calculate Weir and Cockerham's FST from allele frequencies}
\usage{
fstWC_freqs(freqMat, sampMat, doPairs = FALSE, doDist = FALSE,
  perLocus = FALSE)
}
\arguments{
\item{freqMat}{Matrix: Ref allele counts. Rows = populations,
columns = loci; make sure both are named. Row names used to label output FST matrix.}

\item{sampMat}{Matrix: Number of sampled individuals. Rows = populations,
columns = loci.}

\item{doPairs}{Logical: Should pairwise FSTwc be calculated (TRUE) or the
among population FSTwc (FALSE)?}

\item{doDist}{Logical: Should a a distance matrix of FST be returned as well?
Will only run if \code{doPairs==TRUE}.}

\item{perLocus}{Logical: Should the per locus FST be returned? Default is FALSE.}
}
\description{
Takes a matrix of biallelic allele frequencies, and a matrix of sample sizes,
and calculates Weir and Cockerham's FST, i.e. theta (Weir & Cockerham, 1984).
}
\examples{
data(genomalicious_Freqs)
freqMat <- genomalicious_Freqs
sampMat <- matrix(rep(30, 32), nrow=4, ncol=8)
rownames(sampMat) <- paste0('Pop', 1:4)
colnames(sampMat) <- colnames(freqMat); rownames(sampMat) <- rownames(freqMat)

fstWC_freqs(freqMat, sampMat, doPairs=FALSE)
fstWC_freqs(freqMat, sampMat, doPairs=TRUE)
fstWC_freqs(freqMat, sampMat, doPairs=TRUE, doDist=TRUE)

}
\references{
Weir, Cockerham (1984) Estimating F-statistics for the analysis of population structure. Evol. \cr
Weir, Hill (2002) Estimating F-statistics. Annu. Rev. Genet
}
