% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_genos.R
\name{fstWC_genos}
\alias{fstWC_genos}
\title{Calculate Weir and Cockerham's FST from genotypes}
\usage{
fstWC_genos(
  dat,
  popCol = "POP",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT",
  permTest = FALSE,
  bootCI = FALSE,
  iters = 100,
  doPairs = FALSE,
  doDist = FALSE,
  perLocus = FALSE
)
}
\arguments{
\item{dat}{Data table: A long format data table of biallelic genotypes,
coded as '/' separated alleles ('0/0', '0/1', '1/1'), or counts
of the Alt alleles (0, 1, 2, repsectively).
Four columns are required:
\enumerate{
   \item The population ID (see param \code{popCol}).
   \item The sampled individual ID (see param \code{sampCol}).
   \item The locus ID (see param \code{locusCol}).
   \item The genotype (see param \code{genoCol}).
}}

\item{popCol}{Character: The column name with the population information.
Default = \code{'POP'}.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default = \code{'GT'}.}

\item{permTest}{Logical: Should a permutation test be conducted?
Default = \code{FALSE}.}

\item{bootCI}{Logical: Should bootstrap confidence intervals be estimated?
Default = \code{FALSE}.}

\item{iters}{Integer: The number of permutations or bootstrap replicates.
Default = 100.}

\item{doPairs}{Logical: Should pairwise FST be calculated? Default = \code{FALSE},
which calculates the among population FST.}

\item{doDist}{Logical: Should a a distance matrix of FST be returned?
Default = \code{FALSE}. Only applied when \code{doPairs==TRUE}.}

\item{perLocus}{Logical: Should the per locus FST be returned?
Default = \code{FALSE}.}
}
\value{
Returns a list, the contents vary depending on argument choices. \cr
When \code{doPairs==FALSE}, the analysis is among populations.
\enumerate{
   \item \code{$multilocus.mean}: the multilocus FST.
   \item \code{$multilocus.perms}: a vector of permuted FST values.
   \item \code{$multilocus.pval}: the p-value from permutation tests.
   \item \code{$multilocus.ci}: a vector of bootstrap confidence intervals.
   \item \code{$perlocus}: a data table of the FST for each locus.
}
When \code{doPairs==TRUE}, the analysis is between population pairs.
\enumerate{
   \item \code{$multilocus.mean}: a data table of multilocus FST.
   \item \code{$multilocus.ci}: a data table of bootstrap confidence intervals.
   \item \code{$multilocus.dist}: a distance matrix of multilocus FST values.
   \item \code{$perlocus}: a data table of the FST for each locus.
}
}
\description{
Function takes a long format genotype data table and calculates
Weir & Cockerhams FST, i.e. theta (Weir & Cockerham, 1984). Can
also conduct permutation testing and calculate bootstrap
confidence intervals.
}
\details{
Permutation tests involve random shuffling of individuals
among populations, recalculating FST, and testing the hypothesis
that the permuted FST > observed FST. The p-value represents the
proportion of iterations that were "TRUE" to this expression. \cr \cr
Bootstrapping randomly samples the loci with replacement
to estimate the distribution of FST values around the multilocus
FST. Confidence intervals are calculated as the 2.5 and
97.5 percentiles of the obtained bootstrapped FST values.
}
\examples{
data(data_4pops)

# Among popuation FST, with permutation tests
# and bootstrap confidence intervals.
fst_among <- fstWC_genos(data_4pops
                , permTest=TRUE
                , bootCI=TRUE
                , iters=50)

# Pairwise FST, with per locus estimates
# and a distance matrix
fst_pairs <- fstWC_genos(data_4pops
                , doPairs=TRUE
                , doDist=TRUE
                , perLocus=TRUE)

}
