% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_perm.R
\name{fstWC_perm}
\alias{fstWC_perm}
\title{Permutated Weir & Cockerham's FST}
\usage{
fstWC_perm(genoMat, num.cores = 1, pop_id, perms = 100)
}
\arguments{
\item{genoMat}{Matrix: A genotype matrix with samples in rows, loci in
columns, and Alt allele counts in cells (biallelic only: 0, 1, 2).}

\item{num.cores}{Integer: Number of cores. Default is 1.}

\item{pop_id}{Charater: A vector of population IDs. Must be the same length
as \code{nrow(genoMat)} and the order of values must match the order
of rows in \code{genoMat}.}

\item{perms}{Integer: The number of permutations to conduct. Default = 100.}
}
\value{
Returns a vector of permuted multilocus FST values.
}
\description{
Generate a vector of permuted Weir and Cockerham's FST from observed
genotype data.
}
\references{
Weir, Cockerham (1984) Estimating F-statistics for the analysis of population structure. Evol. \cr
}
