% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_permgenos.R
\name{fstWC_permgenos}
\alias{fstWC_permgenos}
\title{Permutation significance test for Weir & Cockerham's FST}
\usage{
fstWC_permgenos(
  dat,
  popCol = "POP",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT",
  perms = 100
)
}
\arguments{
\item{dat}{Data table: A long format data table of biallelic genotypes,
coded as '/' separated alleles ('0/0', '0/1', '1/1'), or counts
of the Alt alleles (0, 1, 2, repsectively).
Four columns are required:
\enumerate{
   \item The population ID (see param \code{popCol}).
   \item The sampled individual ID (see param \code{sampCol}).
   \item The locus ID (see param \code{locusCol}).
   \item The genotype (see param \code{genoCol}).
}}

\item{popCol}{Character: The column name with the population information.
Default = \code{'POP'}.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default = \code{'GT'}.}

\item{perms}{Integer: The number of permutations to conduct. Default = 100.}
}
\value{
Returns a vector of permuted mean FST values (estimated across
all loci).
}
\description{
Tests whether an observed FST value is greater than what could
be observed if samples were randomly permuted among populations.
}
\references{
Weir, Cockerham (1984) Estimating F-statistics for the analysis of population structure. Evol. \cr
Weir, Hill (2002) Estimating F-statistics. Annu. Rev. Genet.
}
