% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_varcomps.R
\name{fstWC_varcomps}
\alias{fstWC_varcomps}
\title{Calculate the variance components of Weir & Cockerham's FST}
\usage{
fstWC_varcomps(freqMat, sampMat)
}
\arguments{
\item{freqMat}{Matrix: Ref allele counts. Rows = populations,
columns = loci; make sure both are named. Row names used to label output FST matrix.}

\item{sampMat}{Matrix: Number of sampled individuals. Rows = populations,
columns = loci.}
}
\description{
Calculate the variance components of Weir & Cockerham's FST
}
