% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_resamp_sim_phenos.R
\name{gam_resamp_sim_phenos}
\alias{gam_resamp_sim_phenos}
\title{Resampling-simulation framework for genomic animal models: Phenotypes}
\usage{
gam_resamp_sim_phenos(
  snpSetList,
  genosDT,
  runsPerSet,
  seqType,
  qtlNumber,
  h2Levels,
  effectSizes,
  numCores = 1
)
}
\arguments{
\item{snpSetList}{List: The output from the function \code{gam_resamp_sim_snps}.}

\item{genosDT}{Data.table: The observed genotype data for each individual.
Requires the columns:
\enumerate{
\item \code{$LINK.BLOCK}: The linkage block ID.
\item \code{$LOCUS}: The locus ID.
\item \code{$SAMPLE}: The sample ID.
\item \code{$GT}: The genotype score (0, 1, or 2)
}}

\item{runsPerSet}{Integer: The number of simulation runs to perform for each
SNP set. Recommend a minimum of 3-5. Each simulation will randomly draw a
`qtlNumber` of loci to act as QTLs. So you can think of this argument as
determining the number of resampled nested QTL sets to simulate within each
SNP set.}

\item{seqType}{Character: The type of sequencing data used. One of
\code{'wgs'} for whole-genome sequencing (linkage blocks are genomic windows),
or \code{'rrs'} for reduced-representation sequencing (linkage blocks are
contigs, or genomic fragments).}

\item{qtlNumber}{Integer: The number of loci to draw as simulated QTLs.}

\item{h2Levels}{Numeric: A vector of heritability levels to simulate. Must be
between 0 and 1.}

\item{effectSizes}{Numeric: A vector of QTL effect sizes to draw from in
simulations. All values must be positive. If you want to simulate a uniform
effect size, simply specify a single value of 1. However, you can use a distribution
of values if you want to simulate variable effect sizes across QTLs.}

\item{numCores}{Integer: Number of cores to run simulations in parallel.}
}
\value{
Returns a list with one slot for each SNP set. Each slot contains
a data.table with the following columns:
\enumerate{
\item \code{$AMAT}: The A-matrix scenario.
\item \code{$SET}: The SNP set ID.
\item \code{$HERIT}: The simulated heritability level.
\item \code{$RUN}: The simulated run ID.
\item \code{$SAMPLE}: The sample ID.
\item \code{$G}: The genetic value.
\item \code{$E}: The residual environmental value.
\item \code{$P}: The phenotypic value (G+E).
}
}
\description{
This is one of a set of functions that support use of a resampling-simulation
framework for assessing the robustness of genomic animal models. These functions
form a pipeline:
\enumerate{
\item \code{gam_resamp_sim_snps}
\item \code{gam_resamp_sim_phenos}
\item \code{gam_resamp_sim_stats}
\item \code{gam_resamp_sim_plot}
}
The \code{gam_resamp_sim_phenos} function is used to generate a suite of
phenotypes for different SNP sets to explore how different compositions of
genetic markers influence the performance of genomics animal models.
}
\details{
This function is effectively calling \code{family_sim_qtl} under the
hood. If you want more information about that function, see \code{?family_sim_qtl}.

Two scenarios are simulated with respect to whether QTLs will be present
in the calculated A-matrix (genomic relationship matrix): 'with QTLs' and 'without QTLs'.

In the 'with QTLs' scenario, loci in the focal SNP set are randomly drawn to
use as simulated QTLs. This represents a 'best case' context where the focal
SNP set captures all the loci contributing to a traits additive genetic variance.

In the contrasting scenario, 'without QTLs', loci that are not in the focal
SNP set are randomly drawn to simulate as QTLs. For \code{seqType=='wgs'},
these can be any loci found throughout the genome that is not in the focal SNP set.
For \code{seqType=='rrs'}, loci in the reserved contigs are are used, and one
locus is drawn per contig. Therefore, if \code{seqType=='rrs'}, you must make
sure that \code{qtlNumber} is the same as the number of reserved SNP sets.
}
