% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_resamp_sim_plot.R
\name{gam_resamp_sim_plot}
\alias{gam_resamp_sim_plot}
\title{Resampling-simulation framework for genomic animal models: Summary plot}
\usage{
gam_resamp_sim_plot(
  obsEstimDT,
  simEstimDT,
  xLabel = "Estimate",
  yLabel = "Number of SNP sets",
  AmtxColours = NULL,
  boxYpos = NULL,
  boxWidth = 1.5
)
}
\arguments{
\item{obsEstimDT}{Data.table: The observed estimates across SNP sets.
Requires the columns:
\enumerate{
\item \code{$SET}: The SNP set ID.
\item \code{$ESTIM}: The estimated value.
}}

\item{simEstimDT}{Data.table: The simulated estimates across SNP sets. Note,
for each SNP set, you will need to take the median across nested runs.
Requires the columns:
\enumerate{
\item \code{$AMAT}: The A-matrix scenario.
\item \code{$SET}: The SNP set ID.
\item \code{$HERIT}: The simulated heritability level.
\item \code{$ESTIM}: The estimated value.
}}

\item{xLabel}{Character: The label for the x-axis. Default is 'Estimate'.}

\item{yLabel}{Character: The label for the y-axis. Default is 'Number of SNP sets'.}

\item{AmtxColours}{Character: A vector of three colours for plotting the A-matrix
scenarios. Values are in this order: 'Observed', 'Simulated (without QTLs)',
'Simulated (with QTLs)'. Default is NULL, so if unspecified, automatic colours
will be used.}

\item{boxYpos}{Numeric: A vector of three values for positioning boxes for boxplots
below underneath histograms. These values must be negative. Values are in this
order: 'Observed', 'Simulated (without QTLs)', 'Simulated (with QTLs)'.
Default is NULL, so if unspecified, automatic values will be used.}

\item{boxWith}{Numeric: A single value, the width of boxes for boxplots.
Default is 1.5.}
}
\value{
Returns a ggplot object. For each heritability level simulated,
the observed and simulated estimates will be illustrated with histograms,
with boxplots summarising the distribution below them.
}
\description{
This is one of a set of functions that support use of a resampling-simulation
framework for assessing the robustness of genomic animal models. These functions
form a pipeline:
\enumerate{
\item \code{gam_resamp_sim_snps}
\item \code{gam_resamp_sim_phenos}
\item \code{gam_resamp_sim_stats}
\item \code{gam_resamp_sim_plot}
}
The \code{gam_resamp_sim_plot} function is used to generate a summary plot
of the observed and simulated estimates. You will need to
compile the estimates from your analyses into data.tables that will be fed
into this function. Estimates could be anything (heritability, additive variance,
residual environmental variance, model comparison P-values, etc).
}
