% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_resamp_sim_snps.R
\name{gam_resamp_sim_snps}
\alias{gam_resamp_sim_snps}
\title{Resampling-simulation framework for genomic animal models: SNP set}
\usage{
gam_resamp_sim_snps(uniqLocusDT, seqType, numSnpSets, reserveBlocksRRS = NULL)
}
\arguments{
\item{seqType}{Character: The type of sequencing data used. One of
\code{'wgs'} for whole-genome sequencing (linkage blocks are genomic windows),
or \code{'rrs'} for reduced-representation sequencing (linkage blocks are
contigs, or genomic fragments).}

\item{numSnpSets}{Integer: The number of unique resampled SNP sets to draw.
One random SNP will be drawn per linkage block.}

\item{reserveBlocksRRS}{Numeric: The number of linkage blocks to reserve for the
for reduced-representation sequencing datasets (\code{seqType=='rrs'}). If you
intend to use the full pipeline, then you need to specify this value to be >0.
These reserved linkage blocks will be used to simulate unsampled genomic regions
in the function, \code{gam_resamp_sim_phenos}. You should \code{reserveBlocksRRS}
to be the maximum number of QTLs you would like to simulate, because the default
behaviour of \code{gam_resamp_sim_phenos} is to draw a maximum of one locus
per reserved linkage block.}
}
\value{
Returns a list with \code{numSnpSets} number of slots. Each slot
contains the following indexed items:
\enumerate{
\item \code{$focal}: Data.table with the randomly sampled SNPs (per linkage
   block) that will be used in this SNP set, with columns \code{$SET},
   \code{$LINK.BLOCK}, and \code{$LOCUS}. This will be present for both
   \code{seqType == 'wgs'} and \code{seqType == 'rrs'}. For \code{seqType == 'rrs'},
   a subset of \code{reserveBlocksRRS} linkage blocks will be dropped randomly
   and reserved for simulation downstream.
\item \code{$reserved}: Data.table with the reserved linkage blocks for the
   SNP set when \code{seqType == 'rrs'}, with columns \code{$SET} and \code{$LINK.BLOCk}.
}
}
\description{
This is one of a set of functions that support use of a resampling-simulation
framework for assessing the robustness of genomic animal models. These functions
form a pipeline:
\enumerate{
\item \code{gam_resamp_sim_snps}
\item \code{gam_resamp_sim_phenos}
\item \code{gam_resamp_sim_stats}
\item \code{gam_resamp_sim_plot}
}
The \code{gam_resamp_sim_snps} function is used to generate a suite of
resampled SNP sets to explore how different compositions of genetic markers
influence the performance of genomics animal models.
}
