% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_resamp_sim_stats.R
\name{gam_resamp_sim_stats}
\alias{gam_resamp_sim_stats}
\title{Resampling-simulation framework for genomic animal models: Summary statistics}
\usage{
gam_resamp_sim_stats(obsEstimDT, simEstimDT)
}
\arguments{
\item{obsEstimDT}{Data.table: The observed estimates across SNP sets.
Requires the columns:
\enumerate{
\item \code{$SET}: The SNP set ID.
\item \code{$ESTIM}: The estimated value.
}}

\item{simEstimDT}{Data.table: The simulated estimates across SNP sets. Note,
for each SNP set, you will need to take the median across nested runs.
Requires the columns:
\enumerate{
\item \code{$AMAT}: The A-matrix scenario.
\item \code{$SET}: The SNP set ID.
\item \code{$HERIT}: The simulated heritability level.
\item \code{$ESTIM}: The estimated value.
}}
}
\value{

}
\description{
This is one of a set of functions that support use of a resampling-simulation
framework for assessing the robustness of genomic animal models. These functions
form a pipeline:
\enumerate{
\item \code{gam_resamp_sim_snps}
\item \code{gam_resamp_sim_phenos}
\item \code{gam_resamp_sim_stats}
\item \code{gam_resamp_sim_plot}
}
The \code{gam_resamp_sim_stats} function is used to summarise observed and
simulated estimates and produce statistics for reporting. You will need to
compile the estimates from your analyses into data.tables that will be fed
into this function. Estimates could be anything (heritability, additive variance,
residual environmental variance, model comparison P-values, etc).
}
