% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_map_plot.R
\name{gene_map_plot}
\alias{gene_map_plot}
\title{Plot a gene map}
\usage{
gene_map_plot(
  mapDT,
  genome_len = NULL,
  gene_colour = NULL,
  gene_type = c("gene", "rRNA"),
  extra_type = c("tRNA", "D-loop"),
  plot_xmax = genome_len,
  extra_ypos = 3,
  plot_ymax = 5,
  gene_txt_size = 4,
  extra_txt_size = 4,
  font = "Arial",
  gene_border = NA,
  gene_size = 2
)
}
\arguments{
\item{mapDT}{Data.table: genetic feature information. Requires the columns: \cr
\enumerate{
   \item $NAME = Character, the name of the genetic feature.
   \item $TYPE = Character, the type of genetic feature.
   \item $STRAND = Integer, the starnd, either 1 or -1.
   \item $START = Integer, the starting base position.
   \item $END = Integer, the ending base position.
}}

\item{genome_len}{Integer: the genome length. Default = NULL. If unspecified,
will be assigned the final base pair of the last genetic feature in \code{mapDT}.}

\item{gene_colour}{Character: a vector of colours to plot genes. Each item
is a colour, with the gene accessible through \code{names(gene_colour)}.
See Details for parameterisation.}

\item{gene_type}{Character: a vector of values present in \code{mapDT$TYPE}
that will be plotted as large coloured bars. For plotting purposes,
"gene features". Default = \code{c('gene', 'rRNA')}.
See Details for parameterisation.}

\item{extra_type}{Character: a vector of values present in \code{mapDT$TYPE}
that will be plotted as small grey bars. For plotting purposes, "extra features".
Default = \code{c('tRNA', 'D-loop')}. See Details for parameterisation.}

\item{plot_xmax}{Numeric: a single value, the maximum x-axis limit.
Default is the genome length, as per \code{genome_len}.}

\item{extra_ypos}{Numeric: a single value, the starting y-axis position for
extra features. See Details for parameterisation.}

\item{plot_ymax}{Numeric: a single value, the maximum y-axis limit.
See Details for parameterisation.}

\item{gene_txt_size}{Integer: a single value, the size for gene feature labels.
Default is 4.}

\item{extra_txt_size}{Integer: a single value, the size for extra feature labels.
Default is 4.}

\item{font}{Character: a single value, the font family to use.
Default is \code{'Arial'}.}

\item{gene_border}{Character: a single value, the colour for borders around
gene features. Default is NA, no border.}

\item{gene_size}{Numeric: a single value, the thickness of borders around
gene features, if a colour if specified in \code{gene_border}. Default is 1.}
}
\value{
Returns a gg object.
}
\description{
Uses base positions information to make a linear plot of genetic features
to produce a gene map. Separates into "gene features" (plotted as large blocks
on the main chromosome) and "extra features" (plotted as small bars offset from
the main chromosome).
}
\details{
There are two major features plotted, "gene features" and
"extra features". These names are just for convention: gene features are
plotted as large coloured bars in center of the plot on the main "chromosome",
whereas extra features are plotted as small grey bars above/below the
gene features, offset from the main chromosome. Anything could be plotted as
a gene or extra feature, and these are specified through \code{gene_type}
and \code{extra_type}.

The name of the genetic feature being plotted is the value of \code{mapDT$NAME}.
This value is effectively evaluated as a mathematical expression to allow
italics for gene names and mixed formatting in gene names. The internal function
call is the evaluation of values by \code{geom_text(..., parse=TRUE)} and
\code{geom_text_repel(..., parse=TRUE)}.

The value of \code{mapDT$STRAND} dictates the position of the coloured bars.
A value of 1 places "genes" on the top of the genomic strand, whereas a value
of -1 places "genes" below the genomic strand.

The colour of the gene features is specified through \code{gene_colour} as
a named vector. If there are two genes, 'COX1' and 'COX2', specification of
their colours can be done like so: \code{c(COX1='pink', COX2='blue')}.
If colours are not specified, one colour is automatically assigned to each
unique "gene".

The value of \code{extra_ypos} specifies that distance of the extra features
from the gene features. Set larger if things are looking squashed.
Additionally, \code{plot_ymax} sets the maximal plotting area, so set this
value larger if things are not fitting well.
}
\examples{
library(genomalicious)

# Create a link to raw external datasets in genomalicious
genomaliciousExtData <- paste0(find.package('genomalicious'), '/extdata')

# Read in a GENBANK file of the Bathygobius cocosensis mitogenome
gbk.read <- mitoGbk2DT(paste(genomaliciousExtData, 'data_Bcocosensis.gbk', sep='/'))
head(gbk.read)

# Subset out the "CDS" types and plot genes, rRNA, tRNA, and D-loop.
# Rename rRNAs for nicer plotting. Because $NAME is evaluated by the
# expression() function, it is useful to put single quotations around characters
to have them read as characters internally by gene_map_plot().
gbk.read[TYPE!='CDS'] \%>\%
.[NAME=='12S ribosomal RNA', NAME:='12S'] \%>\%
.[NAME=='16S ribosomal RNA', NAME:='16S'] \%>\%
.[, NAME:=paste0("'", NAME, "'")] \%>\%
gene_map_plot(mapDT=., genome_len=16692, extra_txt_size=3)

# Plot just the COX genes and the D-loop as "gene features" with
# custom colours and a border. Again, not the use of single quotes nested in
double quotes, which will match up to the edited gene $NAME column below.
gene.col.vec <- c(
"'COX1'"='royalblue',
"'COX2'"='firebrick3',
"'COX3'"='mediumpurple2',
"'CYTB'"='plum3',
"'D-loop'"='grey40')

# Subset focal genes, add quotes to ensure characters are parsed as characters.
gbk.read[NAME \%in\% c('COX1','COX2','COX3','CYTB','D-loop')] \%>\%
.[, NAME:=paste0("'", NAME, "'")] \%>\%
  gene_map_plot(
    mapDT=., genome_len=16692,
    gene_type=c('gene', 'D-loop'), gene_colour=gene.col.vec,
    extra_type=NULL, gene_border='black')

# It is possible to parse characters without the double quotes, but note how
# the '-' character in 'D-loop' has been parsed as a minus symbol.
gbk.read[NAME \%in\% c('COX1','COX2','COX3','CYTB','D-loop')] \%>\%
  gene_map_plot(
    mapDT=., genome_len=16692, gene_type=c('gene', 'D-loop'),
    extra_type=NULL, gene_border='black')

}
