% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDT2alleleDF.R
\name{genoDT2alleleDF}
\alias{genoDT2alleleDF}
\title{Convert a genotype data table to an allele data frame}
\usage{
genoDT2alleleDF(dat, sampCol = "SAMPLE", locusCol = "LOCUS", genoCol = "GT")
}
\arguments{
\item{dat}{Data.table: Contains samples and loci in rows, with a
separate column endcoding genotypes, where each allele is separated
by a '/'.}

\item{sampCol}{Character: Column with sample information. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: Column with locus information. Default = \code{'LOCUS'}.}

\item{genoCol}{Character: Column with genotype information. Default = \code{'GT'}.}
}
\value{
A data frame in wide format, with a sample column \code{$SAMPLE} and
2 * number of loci additional columns containing alleles for each locus.
For example, \code{$locus1_1} and \code{$locus1_2} will contain the first and
second allele for "locus1".
}
\description{
A long-format data table of genotypes is converted into a wide-format
data frame of alleles, where each locus is represented by two columns:
one for the first allele, the other for the second allele.
}
\examples{
data(data_4pops)

data_4pops

alle_df <- genoDT2alleleDF(data_4pops)

alle_df[1:4, 1:6]

}
