% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomalicious.R
\docType{package}
\name{genomalicious}
\alias{genomalicious}
\title{\code{genomalicious}: A smorgasbord of R functions for population genomic analyses}
\description{
An R package containing a variety of functions useful in importing,
formatting, and manipulating data for population genomic analyses. \cr
}
\author{
Joshua Thia (Github: j-a-thia)
}
\keyword{package}
