% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomalicious_PatchyGTs.R
\docType{data}
\name{genomalicious_PatchyGTs}
\alias{genomalicious_PatchyGTs}
\title{An example of a patchy genotype matrix}
\format{A matrix with 100 rows and 100 columns.}
\usage{
data(genomalicious_PatchyMat)
}
\description{
A matrix matrix of genotypes with missing data.
}
\details{
Samples IDs are in \code{rownames(genomalicious_PatchyGTs)}, and locus
IDs are in \code{colnames(genomalicious_PatchyGTs)}. Missing data are
coded by \code{NA.}
}
\keyword{datasets}
