% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomalicious_PoolReads.R
\docType{data}
\name{genomalicious_PoolReads}
\alias{genomalicious_PoolReads}
\title{An example of pool-seq read info}
\format{A data table with 96 rows and 10 columns.}
\usage{
data(genomalicious_PoolReads)
}
\description{
A dataset containing read count information from a pool-seq experiment with
replicate library preparations. \cr
\cr
Data of this structure would be used as the \code{dat} parameter in the
function \code{poolne_estim_input}, which prepares pooled replicate sequencing
data for analysis by Gautier et al.'s (2013) \code{poolne_estim}.
}
\details{
\itemize{
  \item \code{CHROM} = The chromosome ID.
  \item \code{POS} = The SNP positions.
  \item \code{REF} = The reference allele.
  \item \code{ALT} = The alternate allele.
  \item \code{QUAL} = The genotype quality score.
  \item \code{SAMPLE} = The sample replicate ID: [Pop ID]_[Rep ID], where
  Pop ID takes values of Pop1-4 and Rep ID takes values of Rep1-3.#'
  \item \code{LOCUS} = The locus ID.
  \item \code{DP} = The depth (total reads observed) at the locus.
  \item \code{RO} = The number of Ref read counts at the locus.
  \item \code{AO} = The number of Alt read counts at the locus.
}
}
\references{
Gautier et al. (2013) Estimation of population allele frequencies from
next-generation sequencing data: pool-versus individual-based genotyping.
Molecular Ecology, 22(14), 3766–3779.
}
\keyword{datasets}
