% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomaliciousGenos.R
\docType{data}
\name{genomaliciousGenos}
\alias{genomaliciousGenos}
\title{An example of genotype data}
\format{A data table with 4 columns and 40 rows.}
\usage{
data(genomaliciousGenos)
}
\description{
A dataset of two populations, with 10 individuals each, and two loci. \cr
\cr
}
\details{
\itemize{
  \item \code{$POP} = The population pool ID.
  \item \code{$INDS} = The individual ID, read Ind[ID]_[Pop].
  \item \code{$LOCUS} = The locus ID.
  \item \code{$GENO} = The biallelic genotype; two alleles (Ref = 0; Alt = 1) separated
  with a '/'.
}
}
\keyword{datasets}
