% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genos2freqs.R
\name{genos2freqs}
\alias{genos2freqs}
\title{Generate a matrix of allele frequencies from genotypes}
\usage{
genos2freqs(dat)
}
\arguments{
\item{dat}{Data table: It is expected that there are only two alleles, and therefore, only three possible genotypes:
0/0, 0/1 (or 1/0), and 1/1, where the Ref allele is '0'. This data.table needs the following columns:
\code{POP}, the population ID; \code{SAMPLE}, the individual ID; \code{LOCUS}, the locus ID;
and \code{GT}, the genotype.}
}
\value{
Returns a matrix of allele frequencies for the Ref allele (coded as '0' in the genotype)
}
\description{
Parses a data table of genotypes and returns a matrix of allele frequencies.
}
\examples{
# Import genotype data
data(genomaliciousGenos)
genomaliciousGenos

# Convert to frequency matrix
genos2freqs(genomaliciousGenos)

}
