% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genos2indAlleles.R
\name{genos2indAlleles}
\alias{genos2indAlleles}
\title{Convert a genotype data table to an allele data frame}
\usage{
genos2indAlleles(dat, sampCol = "SAMPLE", locusCol = "LOCUS", genoCol = "GT")
}
\arguments{
\item{dat}{Data.table: Contains samples and loci in rows, with a
separate column endcoding genotypes, where each allele is separated
by a '/'.}

\item{sampCol}{Character: Column with sample information. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: Column with locus information. Default = \code{'LOCUS'}.}

\item{genoCol}{Character: Column with genotype information. Default = \code{'GT'}.}
}
\value{
A data frame in wide format, with a sample column \code{$SAMPLE} and
2 * number of loci additional columns containing alleles for each locus.
For example, \code{$locus1_1} and \code{$locus1_2} will contain the first and
second allele for "locus1".
}
\description{
A long-format data table of genotypes is converted into a wide-format
data table of alleles, where each locus is represented by two columns:
one for the first allele, the other for the second allele.
}
\examples{
library(genomalicious)
data(data_4pops)

data_4pops

alle_dt <- genos2indAlleles(data_4pops)

# The first three loci, with count of allele 1 and 2.
alle_dt[1:4, 1:7]

}
