% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/He_freqs.R
\name{He_freqs}
\alias{He_freqs}
\title{Calculate expected heterozygosity from allele frequencies}
\usage{
He_freqs(dat)
}
\arguments{
\item{dat}{Matrix: Rows are population pools, columns are loci, cells are the Ref allele
frequencies.}
}
\value{
Returns a list with the following indexed data tables: \cr
\enumerate{
    \item \code{poolAv} = The overall pool mean across loci.
    \item \code{locusAv} = The He for each locus and pool combination.
}
For each itemed index, there is a data table with columns \code{$POOL} (= population
pool ID) and \code{$HE} (= the expected heterozygosity). Additionally, in the
index \code{$locusAv}, there is a column \code{$LOCUS} (= the locus ID).
}
\description{
Takes a matrix of allele frequencies for different population pools,
at different loci, and returns various statistics for expected heterozygosity (He).
Assumes biallelic data.
}
\details{
Calculates He as 2 * p * q; where p = the Ref allele frequency, and
q = the Alt allele frequency.
}
\examples{
data(pgposerFreqs)

He <- He_freqs(pgposerFreqs)

}
