% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idip_DT.R
\name{idip_DT}
\alias{idip_DT}
\title{Calculate diversity statistics with \code{IDIP} from a data table of genotypes}
\usage{
idip_DT(snpDat, strucMat, sampCol = NA, locusCol = NA, genoCol = NA)
}
\arguments{
\item{snpDat}{Data table: a long-format data table with SNP genotypes
coded as per VCF specifications, e.g. ('0/0', '0/1', '1/1').
Three columns are required:
\enumerate{
   \item (1) The sampled individual ID (see param \code{sampCol}).
   \item (2) The locus ID (see param \code{locusCol}).
   \item (3) The genotype (see param \code{genoCol}).
}}

\item{sampCol}{Character: The column name with the sampled individual information.}

\item{locusCol}{Character: The column name with the locus information.}

\item{genoCol}{Character: The column name with the genotype information.}
}
\description{
Calculate diversity statistics with \code{IDIP} from a data table of genotypes
}
\examples{
data(genomalicious4pops)
snpDat <- genomalicious4pops
}
