% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idip_DTgenos.R
\name{idip_DTgenos}
\alias{idip_DTgenos}
\title{Calculate diversity statistics with \code{IDIP} from a data table of genotypes}
\usage{
idip_DTgenos(snpDat, strucMat, popCol = "POP", locusCol = "LOCUS",
  genoCol = "GT")
}
\arguments{
\item{snpDat}{Data table: a long-format data table with SNP genotypes
coded as per VCF specifications, e.g. ('0/0', '0/1', '1/1').
Three columns are required:
\enumerate{
   \item (1) The population ID (see param \code{popCol}).
   \item (2) The locus ID (see param \code{locusCol}).
   \item (3) The genotype (see param \code{genoCol}).
}}

\item{strucMat}{Matrix: Hierarchy of "aggregates", minimum of 2 rows and
as many columns as the lowest units of the hierarchy. 1st row is the top most
level (e.g. ecosystem or metapopulation), with lower levels in proceding rows.}

\item{popCol}{Character: The column name with the population information, i.e.
the lowest levels of the population hierarchy (last row of \code{strucMat}).}

\item{locusCol}{Character: The column name with the locus information.}

\item{genoCol}{Character: The column name with the genotype information.}
}
\value{
A matrix with columns representing loci, rows are the statistics
from the \code{IDIP} analysis.
}
\description{
Calculate diversity statistics with \code{IDIP} from a data table of genotypes
}
\examples{
data(genomalicious_4pops)
snpDat <- genomalicious_4pops
strucMat <- matrix(c(rep('metapop', 4)
               , paste('Group', c(1,1,2,2))
               , paste('Pop', 1:4, sep='')), ncol=4, byrow=TRUE)
}
