% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_join_DT.R
\name{kmeans_join_DT}
\alias{kmeans_join_DT}
\title{Join K-means clustering results into a reference (long) \code{data.table}}
\usage{
kmeans_join_DT(refTab, sampCol = "SAMPLE", kFit)
}
\arguments{
\item{refTab}{Data.table: the reference data.table with a column of sample
names, see param \code{sampCol}.}

\item{sampCol}{Character: a column in \code{refTab} where sample names are
stored. These sample names must match those in \code{names(kFit$cluster)},
see param \code{kFit}. Default is \code{SAMPLE}.}

\item{kFit}{Kmeans: an object produced by the function \code{kmeans}.
The cluster predictions in \code{kFit$cluster} will be aligned against samples
in \code{refTab[[sampCol]]}. It is therefore necessary that the names in
this object match those expected in \code{refTab}.}
}
\description{
Takes the results of K-means clustering performed by \code{kmeans} and
joins them into a reference data table including the same samples.
}
\details{
It is important that all the samples in \code{refTab} match the names
of samples in \code{kFit}. The function will throw a warning if this is not
the case, but will force the join regardless.
}
\examples{
library(genomalicious)

data(data_4pops)
data_4pops

# Perform PCA.
PCA <- pca_genos(data_4pops)

# Get a table of PC axes
pcTable <- pca2DT(PCA)

# K-means clustering
K4 <- kmeans(PCA$x[,1:4], centers=4, nstart=10)

K4$cluster

# Add in inferred populations.
pcTable2 <- kmeans_join_DT(pcTable, 'SAMPLE', K4)
pcTable2

# Plot
(ggplot(pcTable2, aes(x=PC1, y=PC2, colour=CLUSTER))
   + geom_point()
)

}
