% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_overlap_2groups.R
\name{locus_overlap_2groups}
\alias{locus_overlap_2groups}
\title{Proability of locus overlap between two groups}
\usage{
locus_overlap_2groups(sigA, sigB, lociA, lociB, perms)
}
\arguments{
\item{sigA}{Character/Integer: A vector of loci in group A.}

\item{sigB}{Character/Integer: A vector of loci in group B.}

\item{lociA}{Character/Integer: A vector of loci analysed in group A.}

\item{lociB}{Character/Integer: A vector of loci analysed in group B.}

\item{perm}{Integer: The number of permutations to run for significance testing.}
}
\value{
The value returned is the proportion of null permutations that were greater or
equal to the observed value (i.e. the empirical number of significant loci shared between groups).
}
\description{
This test returns a probability that sets of "significant" loci
identified in two groups do not overlap by chance. For example,
if outlier tests are carried out in two different groups, the overlap
between these groups can be assessed to determine if the number of outliers
shared is significant or could be observed due to random chance.
}
\details{
The hypothesis being tested is that the number of observed loci overlapping
between two groups is greater than that observed due to random chance. Permutations
are run where loci are randomly drawn from each groups set of loci at a size equivalent
to that deemed statistically significant. In other words, if a test determined 2 loci out
of 100 were significant, a null draw would randomly sample 2 loci from the set of 100.
}
\examples{
locus_overlap_2groups(sigA=c(1,5), sigB=5, lociA=1:100, lociB=1:100, perm=1000)

}
