% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_sim_structure.R
\name{miss_sim_structure}
\alias{miss_sim_structure}
\title{Simulate missing data with a specified strcuture}
\usage{
miss_sim_structure(
  dat_clean,
  miss_loci,
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT"
)
}
\arguments{
\item{dat_clean}{Data table: Must be in long format and contains
no missing data (i.e. it is "clean"). Genotypes can be coded as
'/' separated characters or a integers. Must contain the following
columns:
\enumerate{
   \item The sampled individual ID (see param \code{sampCol}).
   \item The locus ID (see param \code{locusCol}).
   \item The genotype (see param \code{genoCol}).
}}

\item{miss_loci}{Numeric: Vector of proportions. Each value represents the
proportion of missing data at a locus.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default is \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default is \code{'GT'}.}
}
\description{
Takes a long format data table of genotypes that
has no missing data (clean) and simulates missing data following
specified per locus proportions.
}
\details{
The number of values in \code{miss_loci} does not have to be equal to
\code{dat_clean}. In the simulations, the values in \code{miss_loci}
are randomly sampled (with replacement) to create a bootstrap simulated
missing data structure of the same dimensions (loci and samples) as the
observed clean data. This simulated missing data structure is then used to
modify the observed clean data. \cr
\cr
NOTE: This method is blind to population, so will simulate
missing data across individuals. Therefore, if missing data exhibits
population structuring, it would be necessary to subset the clean dataset
by population and run specific simulations per population, then combine
all the simulated subsets.
}
\examples{
library(data.table)
library(tidyverse)
data(data_4pops)

# Create a vector of proportions
prop_miss <- rbeta(1000, 0.5, 10)
hist(prop_miss, main='Input missing proportions')

# Simulate missing data structure
patchy4pops <- miss_sim_structure(
    dat_clean=data_4pops
    , miss_loci=prop_miss
    , sampCol='SAMPLE'
    , locusCol='LOCUS'
    , genoCol='GT'
)

patchy4pops[, sum(GT=='./.')/length(GT), by=LOCUS]$V1 \%>\%
hist(main='Simulated missing distribution')

}
