% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missHeatmap.R
\name{missHeatmap}
\alias{missHeatmap}
\title{Plot a heatmap of missing information}
\usage{
missHeatmap(dat, sampCol = "SAMPLE", locusCol = "LOCUS",
  genoCol = "GT", popCol = NA, plotColours = "white", plotNCol = 2)
}
\arguments{
\item{dat}{Data table: Contains genetic information and must have
the following columns,
\enumerate{
  \item The sampled individuals (see param \code{sampCol}).
  \item The locus ID (see param \code{locusCol}).
  \item The genotype column, e.g. a genotype of allele frequency,
  (see param \code{genoCol}).
}}

\item{sampCol}{Character: The column name with the sampled
individual ID. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus ID.
Default = \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype info.
Default = \code{'GT'}. Missing data should be represeted by \code{NA}.}

\item{popCol}{Character: The column name with the population ID.
Optional parameter. Default = \code{NA}.}

\item{plotColours}{Character: Vector of colours to use in plotting.
Size depends on values specified for parameters \code{type} and
\code{popCol}, see Details.}

\item{plotNCol}{Integer: The number of columns to arrange indiviudal
population plots into. Only takes effect when \code{popCol} is specified.
Default = 2.}
}
\description{
Use to visualise the presence or absence of genetic information
for all sample and locus combinations. Can also be subset
by population.
}
\examples{
####   MISSING GENOTYPE DATA   ####
data(genomalicious_4pops)
datGt <- genomalicious_4pops

# Add missing values
datGt <- do.call('rbind'
                , lapply(split(datGt, datGt$SAMPLE), function(x){
                  if(x$POP[1]=='Pop1'){ pr <- 0.1
                  } else if(x$POP[1]=='Pop2'){ pr <- 0.2
                  } else{ pr <- 0.05}
                  numMiss <- rnbinom(1, size=8, prob=pr)
                  idxMiss <- sample(1:nrow(x), size=numMiss, replace=FALSE)
                  x$GT[idxMiss] <- NA
                 return(x)
                }))

head(datGt, 10)

# Heatmaps, default and specified colours
missHeatmap(datGt)
missHeatmap(datGt, plotColours=c('black', 'plum2'))

# Heatmaps, by population
missHeatmap(datGt, popCol='POP')

####   CATCH PLOT OUTPUT FOR LATER USE   ####
gg4pops <- missHeatmap(datGt, popCol='POP')
plot(gg4pops)

}
