% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoGbk2DT.R
\name{mitoGbk2DT}
\alias{mitoGbk2DT}
\title{Convert a mitogenome GENBANK file into a Data table}
\usage{
mitoGbk2DT(
  genbankFile,
  type_vec = c("gene", "CDS", "tRNA", "rRNA", "D-loop", "misc_feature")
)
}
\arguments{
\item{genbankFile}{Chracter: the path to the GENBANK file.}

\item{type_vec}{Chracter: a vector of features to extract. The default value
has been set to extract the major elements of mitochondrial genomes, but
it could be edited to extract other features. Default is
\code{c('gene', 'CDS', 'tRNA', 'rRNA', 'D-loop')}.}
}
\value{
Returns a data.table with the following columns: \cr
\enumerate{
  \item \code{$NAME} = The name of the genetic feature.
  \item \code{$TYPE} = The type of genetic feature.
  \item \code{$STRAND} = Is given a 1 for the reported strand, and a -1
  if it is on the complementary strand.
  \item \code{$START} = The starting base position.
  \item \code{$END} = The end base position.
}
}
\description{
This function reads in a GENBANK file and outputs an R data.table object.
It has been written for the import of mitogenome data into R, but could be
used for importing other GENBANK files (not rigorously tested).
}
