% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitogenome_seq_repos.R
\name{mitogenome_seq_repos}
\alias{mitogenome_seq_repos}
\title{Reposition a mitogenomic sequence}
\usage{
mitogenome_seq_repos(dnaSeq, newStartPos)
}
\arguments{
\item{dnaSeq}{Character: The mitogenomic sequence.}

\item{newStartPos}{Integer: The new start position for the sequence.}
}
\description{
This function takes a linear sequence of a mitochondrial genome and 
repositions it to start at a new base position.
}
\details{
The sequence is repositioned with respect to the value of 
\code{newStartPos}. For example, consider a value of \code{dnaSeq}: \cr
\cr
     GCGCGCGCGCATATGACTAA \cr
\cr
For a value of \code{newStartPos==10}, the 10th base would become the new
start position and the 9th position in the original sequence is a cutoff.
Below this cutoff, the sequence is translocated to the right. \cr
\cr
     GCGCGCGCG | CATATGACTAA (original sequence, cut at 9th base) \cr\cr
     CATATGACTAAGCGCGCGCG (new sequence, repositioned)
}
