% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes_input.R
\name{mrbayes_input}
\alias{mrbayes_input}
\title{Generate input file for \code{MrBayes}}
\usage{
mrbayes_input(dat.list, type.vec, set.vec, out.file)
}
\arguments{
\item{dat.list}{List: Each index contains a matrix for a data partition.
Could be a matrix of morphological characters, DNA alignments for a gene, etc.
Row names are taxa, column names are the characters, with cells containing
the character states per taxa.}

\item{type.vec}{Character: A vector of the data partition types (the matrices
in \code{dat.list}), as per \code{MrBayes} "datatype" requirements:
\code{"Standard"}, \code{"DNA"}, \code{"RNA"}, \code{"Protein"}, or
\code{"Restriction"}. Note that the length of \code{type.vec} must be the same
as the length (number of data partitions) in \code{dat.list}.}

\item{set.vec}{Character: A vector of the data partition sets (the matrices
in \code{dat.list}), effectively a set of user defined names/IDs for the
data partitions. Used for generating the "charset" specifications in
\code{MrBayes}.}

\item{out.file}{Character: A single value, the name of the NEXUS file to save.}
}
\value{
Writes a NEXUS file.
}
\description{
Takes a list of data matrices and converts into a NEXUS file that can
be read by \code{MrBayes}. Can be used for multiple data partitions and
data types.
}
\details{
Primary functionality is to convert data generated in R into NEXUS.
Calculates data ranges and sets up partitions. Addition of \code{MrBayes}
run commands are not supported.
}
