% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_patterson.R
\name{normalise_patterson}
\alias{normalise_patterson}
\title{Perform the Patterson et al. (2006) normalisation to a genotype matrix}
\usage{
normalise_patterson(dat)
}
\arguments{
\item{dat}{Matrix: Counts of Ref allele. E.g. the C(i,j) table described
in Patterson et al. (2006).}
}
\value{
Returns a matrix with the same deminsions as \code{dat}, but with
genotypes normalised
}
\description{
Perform the Patterson et al. (2006) normalisation to a genotype matrix
}
