% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outflank_input_DTfreqs.R
\name{outflank_input_DTfreqs}
\alias{outflank_input_DTfreqs}
\title{Convert an allele frequency data table (long format) to \code{OutFLANK} input}
\usage{
outflank_input_DTfreqs(
  dat,
  popCol = "POP",
  locusCol = "LOCUS",
  freqCol = "FREQ",
  indsCol = "INDS",
  HoCol = NULL
)
}
\arguments{
\item{dat}{Data table: The biallelic allele frequency data. Requires the
following columns:
\enumerate{
   \item The population ID.
   \item The locus ID.
   \item The allele frequency of the Reference allele.
   \item The number of individuals (sample size) used to estimate allele frequency.
}}

\item{popCol}{Character: The column name with the sampled individual information. Default = \code{'POP'}.}

\item{locusCol}{Character: The column name with the locus information. Default = \code{'LOCUS'}.}

\item{freqCol}{Character: The column name with the allele frequency information. Default = \code{'FREQ'}.}

\item{indsCol}{Character: The column name with the number of individuals (sample size)
used to estimate the allele frequency. Default = \code{'INDS'}.}

\item{HoCol}{Character: An optional argument. The column name with the observed number of
heterozygotes. Affects estimation of FST, see details. Default = \code{NULL}.}
}
\value{
A data frame in the required format for analysis by \code{OutFLANK}.
}
\description{
Converts a long format data table of allele frequency information into
a data frame fit for analysis by OutFLANK (Whitlock & Lotterhos 2015).
}
\details{

}
\examples{
data('data_PoolPi')

freqs <- data_PoolPi

ofDF <- outflank_input_DTfreqs(freqs, 'POOL', 'LOCUS', 'PI', 'INDS')

}
\references{
Whitlock & Lotterhos (2015) Reliable detection of loci responsible
for local adaptation: Inference of a null model through trimming the distribution
of FST. The American Naturalist 186.
}
