% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outflank_mod_fst_nocorrect.R
\name{outflank_mod_fst_nocorrect}
\alias{outflank_mod_fst_nocorrect}
\title{Modificatin of OutFLANK's \code{WC_FST_FiniteSample_Diploids_2Alleles_NoCorr()}}
\usage{
outflank_mod_fst_nocorrect(alleleFreqs, sampSize, Ho)
}
\arguments{
\item{alleleFreqs}{Numeric: A vector of the Reference allele frequency, with
the length == number of populations.}

\item{sampSize}{Integer: A vector of sample sizes, with the length == number
of populations and the order corresponding to the those in \code{alleleFreqs}.}

\item{Ho}{Integer: An optional vector of the heterozygote frequency,
with the length == number of populations and the order corresponding 
to the those in \code{alleleFreqs}. Default = \code{NULL}, see Details.}
}
\value{
Returns a list of values related to FST:
 \itemize{
 \item   He:  the expected heterozygosity of the locus
 \item 	FSTNoCorr:  Fst (without sample size correction)
 \item 	T1NoCorr: The numerator of the uncorrected sample size correction (similar to Weir and Cockerham 1984)
 \item   T2NoCorr: The denominator of the uncorrected sample size correction
 }
}
\description{
Skip the need for diploid genotypes, i.e. just use allele frequencies,
and produce the same output as \code{WC_FST_FiniteSample_Diploids_2Alleles_NoCorr()}.
Calculate FST without a sample size correction.
}
\details{
By using just allele frequencies, there is no information about 
the observed frequency of heterozygotes, Ho. The function will make the
naive assumption that Ho = He, if Ho is not specified.
}
