% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseMat2DT.R
\name{pairwiseMat2DT}
\alias{pairwiseMat2DT}
\title{Convert a pairwise matrix into a pairwise data table}
\usage{
pairwiseMat2DT(dat, flip = FALSE, X1, X2, Y)
}
\arguments{
\item{dat}{Matrix or Data table: The default parameterisation (\code{flip==FALSE})
is expecting a pairwise matrix. This matrix needs to be symmetrical, with a
zero on the diagonal and the off-diagonal containing the measured variable
values between pairs. There should be row and column names. If instead you
want to go in the flipped parameterisation (\code{flip==TRUE}), then a
data table is expected, where pair combinations are represented into two
columns and the measured variable is in a third column.}

\item{flip}{Logical: Should the function be flipped? Default is \code{FALSE},
a pairwise matrix should be converted into a pairwise data table. If set to
\code{TRUE}, then a pairwise data table is converted into a matrix.}

\item{X1}{Character: The column name for the first member of the pair.
If \code{flip==FALSE}, then \code{X1} will take the row names from the matrix
and put them into the first column of the output data table.
If \code{flip==TRUE}, then \code{X1} will be take the column in the data table and
put it as row names in the output matrix.}

\item{X2}{Character: The column name for the first member of the pair.
If \code{flip==FALSE}, then \code{X2} will take the column names from the matrix
and put them into the second column of the output data table.
If \code{flip==TRUE}, then \code{X2} will be take the column in the data table and
put it as column names in the output matrix.}

\item{Y}{Character: The column name or the measured variable between pairs.
If \code{flip==FALSE}, then \code{Y} will take the values in the off-diagonal
in the matrix and put them into the third column of the output data table.
If \code{flip==TRUE}, then \code{Y} will be take the column in the data table and
put it as values in the off-diagonal in the output matrix.}
}
\description{
This function takes pairwise matrix (all possible combinations in rows and
columns) and converts it into a long-format data table. This function can
also be flipped, i.e., going from a pairwise data table to a pairwise matrix.
}
\examples{
library(genomalicious)
data(data_Genos)

# Calculate pairwise FST
pairFst <- fstat_calc(data_Genos, type='genos', fstat='FST', global=FALSE, pairwise=TRUE)

# Convert into a pairwise matrix
pairMat <- pairwiseMat2DT(
   dat=pairFst$genome, flip=TRUE, X1='POP1', X2='POP2', Y='FST'
)
pairMat

# Convert the pairwise matrix into a data table
pairDT <- pairwiseMat2DT(
   dat=pairMat, flip=FALSE, X1='POP1', X2='POP2', Y='FST'
)
pairDT

}
