% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_DTfreqs.R
\name{pca_DTfreqs}
\alias{pca_DTfreqs}
\title{Conduct a PCA on allele frequences stored in a (long) data table}
\usage{
pca_DTfreqs(dat, popCol = "POP", locusCol = "LOCUS", freqCol = "FREQ")
}
\arguments{
\item{dat}{Data table: A long data table with the following columns:
\enumerate{
   \item (1) The population ID (see param \code{popCol}).
   \item (2) The locus ID (see param \code{locusCol}).
   \item (3) The Ref allele frequency (see param \code{freqCol}).
}}

\item{popCol}{Character: An optional argument. The column name with the
population information. Default is \code{'POP'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{freqCol}{Character: The column name with the genotype information.
Default is \code{'GT'}.}
}
\value{
Returns a \code{prcomp} object with an additional index of \code{$pops},
the population ID.
}
\description{
Takes a long data table of biallelic population allele frequencies
and conducts a PCA using R's \code{prcomp()} function.
Allele frequencies are scaled to a mean = 0 and variance = 1.
}
\examples{
data(genomaliciousFreqsLong)

# Conduct the PCA
pca <- pca_DTfreqs(genomaliciousFreqsLong)

# The PCA scores
pca$x

# The populations, i.e. rows in pca$x
pca$pops

}
\references{
Patterson et al. (2006) Population structure and eigenanalysis. PLOS Genetics.
}
