% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_DTinds.R
\name{pca_DTinds}
\alias{pca_DTinds}
\title{Conduct a PCA on individual genotypes stored in a data table}
\usage{
pca_DTinds(dat, scaling = "center", sampCol = "SAMPLE",
  locusCol = "LOCUS", genoCol = "GT")
}
\arguments{
\item{dat}{Data table: A long data table, e.g. like that imported from
\code{vcf2DT}.}

\item{scaling}{Character: How should the data (loci) be scaled? Set to
\code{'center'} (default) to scale to mean = 0 and variance = 1. Set to
\code{'patterson'} to use the Patteron et al. (2006) normalisation.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default is \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default is \code{'GT'}.}
}
\description{
Conduct a PCA on individual genotypes stored in a data table
}
