% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_scatter.R
\name{pca_scatter}
\alias{pca_scatter}
\title{Create a scatter plot of a PCA on genotypes}
\usage{
pca_scatter(dat, axisIndex = c(1, 2), pops = NULL, popCols = NULL,
  look = "ggplot")
}
\description{
Create a scatter plot of a PCA on genotypes
}
