% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_freqs.R
\name{pcoa_freqs}
\alias{pcoa_freqs}
\title{Perform a PCoA (principal coordinates analysis) on population allele frequencies}
\usage{
pcoa_freqs(
  dat,
  popCol = "POP",
  locusCol = "LOCUS",
  freqCol = "FREQ",
  indsCol = "INDS"
)
}
\arguments{
\item{dat}{Data table: A long-format data table containing allele
frequency estimates for biallelic SNP loci. Requires the following columns:
\enumerate{
  \item The population ID (see param \code{popCol}).
  \item The locus ID (see param \code{locusCol}).
  \item The allele frequency column column (see param \code{genoCol}).
  \item The number of pooled diploids used to generate the allele frequency
  estimates (see param \code{indsCol})
}}

\item{popCol}{Character: The column name with the population information.
Default is \code{'POP'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{freqCol}{Character: The column name with the allele freuqency information.
Default is \code{'FREQ'}.}

\item{indsCol}{Character: The column name with the number of individuals
contributing to the allele freuqency estimate. Default is \code{indsCol}.}
}
\value{
Returns a list-like object of class \code{eigen}, as per \code{ape::pcoa},
with an additional long-format data table of genome-wide pairiwse FST
estimates in an index \code{$fst}.
}
\description{
Takes a long-format data table of allele frequencies, calculates pairwise FST,
and conducts a PCoA of these pairwise FST values using R's \code{ape::pcoa()}
function (be sure to cite Paradis & Schliep 2019).
}
\examples{
library(genomalicious)
data(data_PoolFreqs)
data(data_PoolInfo)

# Note columns in data_PoolFreqs
colnames(data_PoolFreqs)

# We need to add in the number of diploid individuals, $INDS
newFreqData <- left_join(data_PoolFreqs, data_PoolInfo)
head(newFreqData)

# Fit the PCoA
PCOA <-pcoa_freqs(newFreqData)

# Object class
class(PCOA)

# Axes
PCOA$vectors

# Eigenvalues
PCOA$values

# Pairwise FST in long-format data table
PCOA$fst

}
\references{
Paradis & Schliep (2019) Bioinformatics. DOI: 10.1093/bioinformatics/bty633
}
