% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_plot.R
\name{pcoa_plot}
\alias{pcoa_plot}
\title{Plot PCoA results}
\usage{
pcoa_plot(
  pcoaObj,
  type = "scatter",
  axisIndex = NULL,
  plotColours = NULL,
  plotLook = "ggplot",
  legendPos = "top"
)
}
\arguments{
\item{pcoaObj}{Prcomp object: A PCA of genotype data fitted using the
\code{prcomp} function. Either manually fitted, or using \code{genomalicious::pca_genos}.}

\item{type}{Character: What type of plot to make: a scatterplot (\code{'scatter'}),
a screeplot of explained variances (\code{'scree'}), or the cumulative explained
variance (\code{'cumvar'}).}

\item{axisIndex}{Integer: The PC axes to plot. If \code{type=='scatter'},
then must be exactly 2 values, the two PC axes to plot as a scatterplot.
If either \code{type=='scree'} or \code{type=='cumvar'}, then can be of
length from 1 to p, where p is the number of PC axes, and values again
represent the desired PC axes to plot.}

\item{plotColours}{Character: A vector of colours to use for plotting,
but is an optional argument. Default = \code{NULL}.
When \code{type=='scatter'}, this must be a named list with one colour
per population. When \code{type=='scree'} or \code{type=='cumvar'}, only
a single colour is required, which is the colour of bars in the screeplot
or cumulative variance plot, respectively, and will default to 'grey20'
if unspecified.}

\item{plotLook}{Character: The look of the plot. Default = \code{'ggplot'}, the
typical gray background with gridlines produced by \code{ggplot2}. Alternatively,
when set to \code{'classic'}, produces a base R style plot.}

\item{legendPos}{Character: Where should the legend be positioned? Default is
\code{'top'}, but could also be one of, \code{'right'}, \code{'bottom'},
\code{'left'}, or \code{'none'}.}

\item{pops}{Character: A vector of population IDs, should match the
rows in \code{pcoaObj$x}, but is an optional argument. Default = \code{NULL}.
The function will search for \code{pcoaObj$pops} to assign to this argument
if not specified. Only valid when \code{type=='scatter'}.}
}
\value{
Returns a ggplot object.
}
\description{
Plots results of a PCoA, e.g., scatterplot, screeplot, and cumulative
explained variance plots. Takes \code{pcoa} object as the main input.
}
\examples{
library(genomalicious)
data(data_PoolFreqs)
data(data_PoolInfo)

# Note columns in data_PoolFreqs
colnames(data_PoolFreqs)

# We need to add in the number of diploid individuals, $INDS
newFreqData <- left_join(data_PoolFreqs, data_PoolInfo)
head(newFreqData)

# Fit the PCoA
PCOA <-pcoa_freqs(newFreqData)

# Plot scatter with default settings
pcoa_plot(PCOA, type='scatter')

# Plot scatter with custom colours and a classic look
pcoa_plot(
   PCOA,
   type='scatter',
   plotColours=c(Pop1='gray30', Pop2='royalblue', Pop3='palevioletred3', Pop4='plum2'),
   plotLook='classic'
)

# Explained variance
pcoa_plot(PCOA, type='scree')

# Cumulative variance with custom colour
pcoa_plot(PCOA, type='cumvar', plotColours='royalblue')

}
