% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgposer.R
\docType{package}
\name{pgposer}
\alias{pgposer}
\alias{pgposer-package}
\title{\code{pgposer}: Population genetic of pooled sequences in R.}
\description{
A package for basic population genetic analyses in R. \cr
A main functionality of this package is the use of replicate library preparations
to estimate uncertainty in population genetic inferences.
}
\author{
Joshua Thia: pgposer@gmail.com
}
\keyword{package}
