% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgposerFreqs.R
\docType{data}
\name{pgposerFreqs}
\alias{pgposerFreqs}
\title{An example of population allele frequency data}
\format{A data table with 5 columns and 3 rows.}
\usage{
data(pgposerFreqs)
}
\description{
A dataset containing population allele frequencies. \cr
\cr
Data of this structure would be used as the \code{dat} parameter in the
functions: \cr
\itemize{
  \item \code{Fst_freqs_demes}
  \item \code{Fst_freqs_pairs}
}
}
\details{
Each column is a locus and rows are populations. Use \code{colnames(pgposerFreqs)} and
\code{rownames(pgposerFerqs)} to get the locus and population names, respectively. Cells contain
the frequencies for the Ref allele.
}
\keyword{datasets}
