% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missBYsamps.R
\name{plot_missBYsamps}
\alias{plot_missBYsamps}
\title{Plot missing data, by samples}
\usage{
plot_missBYsamps(dat, type = "hist", look = "ggplot", sampCol = "SAMPLE",
  locusCol = "LOCUS", respCol = "GT", popCol = NA,
  plotColours = "white", plotNCol = 2)
}
\arguments{
\item{dat}{Data table: Contains genetic information and must have
the following columns,
\enumerate{
  \item (1) The sampled individuals (see param \code{sampCOl}).
  \item (2) The locus ID (see param \code{locusCol}).
  \item (3) The response column, e.g. a genotype of allele frequency,
  (see param \code{respCol}).
}}

\item{type}{Character: The type of plot to make. A histogram depicting
the frequency distribution of samples missing information at each
locus (\code{'hist'}, the default)? Or a heatmap illustrating the
missing loci for each sample (\code{'heatmap'})?}

\item{look}{Character: The look of the plot. Default = \code{'ggplot'}, the
typical gray background with gridlines produced by \code{ggplot2}. Alternatively,
when set to \code{'classic'}, produces a base R style plot.}

\item{sampCol}{Character: The column name with the sampled
individual information. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{respCol}{Character: The column name with the response information.
Default = \code{'GT'}. Missing data should be represeted by \code{NA}.}

\item{popCol}{Character: The column name with the population information.
Optional parameter. Default = \code{NA}.}

\item{plotColours}{Character: Vector of colours to use in plotting.
Size depends on values specified for parameters \code{type} and
\code{popCol}, see Details.}

\item{plotNCol}{Integer: The number of columns to arrange indiviudal
population plots into. Only takes effect when \code{popCol} is specified.
Default = 2.}
}
\description{
Use to visualise missing data with respect to samples and
their associated populations.
}
\details{
1) Return a plot, 2) display plot 3) Break down the different
plot types and population inclusions
}
\examples{
data(genomalicious_4pops)
dat <- genomalicious_4pops

# Add missing values.
missIdx <- c(sample(1:nrow(dat), size=0.05*nrow(dat), replace=FALSE)
             , 100:500, 800:1200, 8000:8888, 200000:200500)
dat$GT[missIdx] <- NA

}
