% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyrad_DT2RADdata.R
\name{polyrad_DT2RADdata}
\alias{polyrad_DT2RADdata}
\title{Convert a long-format genotype data table into polyRAD's RADdata class}
\usage{
polyrad_DT2RADdata(
  dat,
  chromCol = "CHROM",
  posCol = "POS",
  refCol = "REF",
  altCol = "ALT",
  roCol = "RO",
  aoCol = "AO",
  sampCol = "SAMPLE"
)
}
\arguments{
\item{dat}{Data.table: A long-format data.table with columns specified in
\code{chromCol}, \code{posCol}, \code{refCol}, \code{altCol}, \code{locusCol},
and in \code{sampCol}.}

\item{chromCol}{Character: Column with chromosome (or contig) information. Default = \code{'CHROM'}.}

\item{posCol}{Character: Column with position information. Default = \code{'POS'}.}

\item{refCol}{Character: Column with reference allele information. Default = \code{'REF'}}

\item{altCol}{Character: Column with alternate allele information. Default = \code{'ALT'}}

\item{roCol}{Character: Column with reference observed read counts information. Default = \code{'RO'}}

\item{aoCol}{Character: Column with alternate observed read information. Default = \code{'AO'}}

\item{sampCol}{Character: Column with sample information. Default = \code{'SAMPLE'}.}
}
\value{
Returns a RADdata object class.
}
\description{
For a data table with samples and loci as rows (long-format data table),
create a RADdata object as required for the R package, polyRAD (Clark et al 2019).
}
\examples{
data(data_Genos)

genoRADdata <- polyrad_DT2RADdata(data_Genos)

genoRADdata

}
\references{
Clark et al (2019). polyRAD: Genotype calling with uncertainty from sequencing data in polyploids and diploids. G3. doi.org/10.1534/g3.118.200913
}
