% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolfstat_boot.R
\name{poolfstat_boot}
\alias{poolfstat_boot}
\title{Bootstrap FST values from \code{poolfstat}}
\usage{
poolfstat_boot(dat, num.boots = 100, method = "Anova", snp.index = NA)
}
\arguments{
\item{dat}{Pooldata: The main data class object for \code{poolfstat}. Can be created
from a data table of read counts using \code{poolfstat_DT}.}

\item{num.boots}{Integer: The number of bootstrap simulations to run. Default = 100.}
}
\value{
Returns a vector of FST values produced by bootstrapping loci in the
original read count dataset, \code{dat}.
}
\description{
Takes a data table of read counts and bootstraps the estimate of FST
calcualted from the function \code{poolfstat::computeFST}. Also requires
pool size information.
}
\examples{
#' # Load in the pool metadata and reads
data(genomaliciousReads)
data(genomaliciousInfo)

# Subset to keep only Rep1 reads.
X <- genomaliciousReads[grep(pattern='Rep1', x=genomaliciousReads$SAMPLE)]

# Need to add pool ID.
X$POOL <- unlist(lapply(strsplit(X$SAMPLE, '_'), function(X){ return(X[1]) }))

# Use poolfstat_DT to compute FST for this dataset and create a pooldata object
poolFst <- poolfstat_DT(X, genomaliciousInfo)

# Bootstrap FST, using pooldata object from poolFst
bootFst <- poolfstat_boot(poolFst$pooldat, 100)

}
