% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolfstat_pooldata2DT.R
\name{poolfstat_pooldata2DT}
\alias{poolfstat_pooldata2DT}
\title{Convert a \code{pooldata} object to a data table of read counts}
\usage{
poolfstat_pooldata2DT(pooldata)
}
\arguments{
\item{pooldata}{Pool data: an \code{pooldata} object, such as that obtained
from importing VCF with \code{poolfstat::vcf2pooldata}.}
}
\value{
Returns a long-format data table with the columns:
\enumerate{
   \item \code{$CHROM} The chromosome (contig) ID.
   \item \code{$POS} The variant position on the chromosome.
   \item \code{$REF} The reference allele.
   \item \code{$ALT} The alternate allele.
   \item \code{$LOCUS} The locus ID ([chrom]_[pos])
   \item \code{$POOL} The pool ID.
   \item \code{$DP} The total number of reads.
   \item \code{$AO} The number of reads supporting the alternate allele.
   \item \code{$RO} The number of reads supporting the reference allele.
}
}
\description{
Takes a \code{pooldata} class object (from \code{poolfstat} package) and
converts into a long-format data.table.
}
\examples{
# Load in the pool metadata and reads
data(data_PoolInfo)
data(data_PoolReads)

# Pool info
data_PoolInfo

# Replicate samples for each pool
data_PoolReads

# Subset to keep only Rep1 reads.
subReads <- data_PoolReads[grep(pattern='Rep1', x=data_PoolReads$SAMPLE)]

# Need to add pool ID: the ID used to match $POOL in data_PoolInfo
subReads$POOL <- unlist(lapply(strsplit(subReads$SAMPLE, '_'), function(X){ return(X[1]) }))

# Make pooldata object
pooldataObj <-poolfstat_DT2pooldata(subReads, data_PoolInfo)

# And go back to data table
pooldataTab <- poolfstat_pooldata2DT(pooldataObj)

}
