% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolne_estim_input.R
\name{poolne_estim_input}
\alias{poolne_estim_input}
\title{Generate input files for \code{poolne_estim}}
\usage{
poolne_estim_input(
  dat,
  pool.info,
  runID,
  mcmc.sampling = c(1000, 25, 5000, 20, 50, sample(as.integer(1000:9999), 1))
)
}
\arguments{
\item{dat}{Data table: The biallelic SNP data. Requires all of the following columns: \cr
\enumerate{
  \item \code{$POOL} = The population pool ID. \cr
  \item \code{$SAMPLE} = The sample replicate ID; e.g. 1, 2, or 3, if three replicates were made. \cr
  \item \code{$CHROM} = The chromosome ID. \cr
  \item \code{$LOCUS} = The locus ID. \cr
  \item \code{$RO} = The number of Ref read counts at the locus. \cr
  \item \code{$AO} = The number of Alt read counts at the locus. \cr
}}

\item{pool.info}{Data table: The population pool metadata. Requires all of the following columns: \cr
\enumerate{
  \item \code{$POOL} = The population pool ID. \cr
  \item \code{$INDS} = The number of diploid individuals in each pooled library.
}}

\item{runID}{Character: Appended inside filename to indicate a particular run or sim identifier.}

\item{mcmc.sampling}{A vector MCMC sampling parameters for poolne_estim. Can be customised, but the
can be left alone (these are the defaults in \code{poolne_estim}). Must have 6 values: \cr
            \enumerate{
               \item The number of values to sample from the posterior distribution (e.g. 1000).
               \item The second Thinning rate (e.g. 25).
               \item The burn-in period length (e.g. 5000).
               \item The maximal number of pilot runs (e.g. 20) to adjust the parameters of the proposal distributions.
               \item The pilot run length (e.g. 500).
               \item A positive interger for random seed generation.
            }}
}
\value{
Four files are written to the working directory: \cr
            \enumerate{
              \item \code{[runID]_[poolID]_Ref.txt} = Counts of Ref alleles; loci in rows, sample replicates in columns.
              \item \code{[runID]_[poolID]_Dep.txt} = Depth of reads; loci in rows, sample replicates in columns.
              \item \code{[runID]_[poolID]_Loci.txt} = A list of loci in the order they appear in the count TXT files.
              \item \code{[runID]_[poolID]_Input.txt} = The params input file for \code{poolne_estim}.
            }
}
\description{
This function prepares biallelic SNP data for analysis by the programme \code{poolne_estim}.
}
\examples{
# Load in the pool metadata and reads
data(data_PoolInfo)
data(data_PoolReads)

# Have a look at the data: the samples are populations
# ('Pop') and replicate library preps ('Rep').
data_PoolReads$SAMPLE

# You need to make sure there is a column that contains
# the pool ID. Split the SAMPLE column and return the first value:
X <- data_PoolReads
X$POOL <- unlist(lapply(strsplit(X$SAMPLE, '_'), function(X){ return(X[1]) }))

# Check
X

# Now make inputs
poolne_estim_input(dat=X, pool.info=data_PoolInfo, runID='genomalicious')

}
