% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads2freqs.R
\name{reads2freqs}
\alias{reads2freqs}
\title{Calculate reference allele frequencies form read counts}
\usage{
reads2freqs(dat)
}
\arguments{
\item{dat}{Data table: The read counts. Requires the following columns: \cr
\enumerate{
   \item \code{$POOL} = the population pool ID.
   \item \code{$LOCUS} = the locus ID.
   \item \code{$DP} = the total read depth.
   \item \code{$RO} = the observed reference read counts.
}}
}
\value{
A matrix of allele freuqencies for each population. Loci in columns, populations
in rows, and Ref allele frequencies in the cells.
}
\description{
Parses a data table of read counts and calculates Ref allele frequencies \cr
as the observed reference reads divided by the total number of reads. This is \cr
primarily meant for pool-seq data.
}
\examples{
data(pgposerReads)

# Pull out a single replicate from the pgposer demo dataset.
rep1Reads <- pgposerReads[grep(pattern='Rep1', x=pgposerReads$SAMPLE),]
rep1Reads

# Add a population pool column:
rep1Reads$POOL <- unlist(lapply(strsplit(rep1Reads$SAMPLE, '_'), function(X){ X[1] }))

# Generate a matrix of Ref allele frequencies
freqMat <- reads2freqs(rep1Reads)
freqMat

}
