% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2DT.R
\name{vcf2DT}
\alias{vcf2DT}
\title{VCF file to data table}
\usage{
vcf2DT(vcfFile)
}
\arguments{
\item{vcfFile}{Character: The path to the input VCF file.}
}
\value{
A data.table with all the columns contained in the original VCF file with
some additions:
\itemize{
    \item A column called \code{LOUCS} is generated. This is the concatenation of the
             \code{CHROM} and \code{POS} column to form a locus ID.
    \item A column called \code{SAMPLE} is generated. This contains the sample IDs that
             are the columns that follow the \code{FORMAT} column in the original VCF.
    \item The items in the original \code{FORMAT} column of the VCF are given their own columns.
} \cr
Note, for VCF files produced by Stacks, the $CHROM is given the same value as the $ID column.
}
\description{
Reads a VCF file and converts to a data table.
}
\examples{
# Create a link to raw external datasets in pgposer
pgposerExtData <- paste0(find.package('pgposer'), '/extdata')

# This command here shows you the VCF file that comes with pgposer
list.files(pgposerExtData, pattern='vcf')

# Use this to create a path to that file
vcfPath <- paste0(pgposerExtData, '/pgposer.vcf.txt')

# You can read the file in as lines to see what it
# looks like:
readLines(vcfPath)

# Now read it in as a data.table
vcf2DT(vcfPath)

}
