% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_sliding_ranges.R
\name{window_sliding_ranges}
\alias{window_sliding_ranges}
\title{Generate genomic sliding window ranges}
\usage{
window_sliding_ranges(dat, step_size, flank_size)
}
\arguments{
\item{dat}{Data.table: The chromosome/contig details. Requires the columns
\code{$CHROM} the sequence ID, the \code{WIDTH} the sequence length in bases.}

\item{step_size}{Numeric: The step size for the sliding windows.}

\item{flank_size}{Numeric: The flank size for the sliding windows.}
}
\description{
This function takes in a data.table of chromosome/contigs details and
returns overlapping window coordinates.
}
\details{
The function aims to generate overlapping sliding windows.
Argument \code{step_size} will dictate the spacing of each window's midpoint.
The width of the window is determined by argument \code{flank_size}, which
extends either side of the midpoint. An error will return if the
\code{flank_size} does not produce overlapping windows.
}
\examples{
library(genomalicious)

chrom.tab <- data.table(CHROM=c('chrom1','chrom2'), WIDTH=c(100000,550000))

slide.ranges <- window_sliding_ranges(chrom.tab, step_size=500, flank_size=400)

print(slide.ranges)
}
