% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_counts.R
\name{allele_counts}
\alias{allele_counts}
\title{Calculate allele counts}
\usage{
allele_counts(dat, biallelic = TRUE)
}
\arguments{
\item{dat}{Character/Integer: A vector of genotypes. If the class
is \code{'character'}, then assumes alleles are separated
with a '/'. Doesn't have to be biallelic (see param \code{biallelic}).
If the class is \code{'integer'}, then assumes counts of
Alt alleles, in which case, it assumes data is biallelic.}

\item{biallelic}{Logical: Is the data biallelic? Affects processing and
output (see also "Value" section).}
}
\value{
A vector of alleles counts. If \code{biallelic==TRUE}, returns
vector with names \code{'ref'} and \code{'alt'}. If \code{biallelic==FALSE},
returns a vector with names as alleles.
}
\description{
Calculate allele counts
}
\examples{
# Genotypes as separated alleles, biallelic
allele_counts(c('1/1', '0/1', '0/1', '0/0', '0/0'))

# Genotypes as separated alleles, not biallelic
allele_counts(c('1/1', '2/3', '1/3', '0/0', '2/2'), biallelic=FALSE)

# Genotypes as counts of the Alt allele
allele_counts(c(2, 1, 1, 0, 0))

}
