% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_counts.R
\name{allele_counts}
\alias{allele_counts}
\title{Calculate allele counts}
\usage{
allele_counts(dat)
}
\arguments{
\item{dat}{Character/Integer: A vector of genotypes. If the class
is \code{'character'}, then assumes alleles are separated
with a '/'. If theclass is \code{'integer'}, then assumes counts of
Ref alleles.}
}
\value{
A vector of alleles counts for Ref ('0') and Alt ('1') alleles.
}
\description{
Calculate allele counts
}
\examples{
allele_counts(c('1/1', '0/1', '0/1', '0/0', '0/0'))
allele_counts(c(2, 1, 1, 0, 2))

}
