% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da2DT.R
\name{da2DT}
\alias{da2DT}
\title{Convert a \code{lda} object to a long-format \code{data.table}}
\usage{
da2DT(daObj, sampVec, obsPops)
}
\arguments{
\item{daObj}{Lda: An object produced by \code{lda}.}

\item{sampVec}{Character: The sample IDs, which need to match the sample
order in \code{predict(daObj)$x}.}

\item{obsPops}{Character/Factor: The observed population IDs, which need to
match sample order in \code{predict(daObj)$x}.}
}
\value{
Returns a wide-format data.table with columns \code{$POP}, the
population IDs, \code{$SAMPLE}, the sample IDs, \code{$LD[x]}, the individual
LD axes, with \code{[x]} denoting the axis number.
}
\description{
Convert a \code{lda} object to a long-format \code{data.table}
}
\examples{
library(genomalicious)

data(data_Genos)

PCA <- pca_genos(data_Genos, popCol='POP')
DA <- lda(PCA$pops ~ PCA$x[,1:3])

da2DT(DA, sampVec=rownames(PCA$x), obsPops=PCA$pops)

}
