% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadi_inputs_pools.R
\name{dadi_inputs_pools}
\alias{dadi_inputs_pools}
\title{Genertate DADI input from pool-seq data}
\usage{
dadi_inputs_pools(dat, poolCol = "POOL", locusCol = "LOCUS",
  refCol = "REF", altCol = "ALT", freqCol = "FREQ", indsCol = "INDS",
  poolSub = NULL)
}
\arguments{
\item{dat}{Data table: Must contain columns with the following information,
\enumerate{
\item Population pool ID
\item Locus ID
\item Reference allele
\item Alternate alelle
\item Reference allele freuqency
\item Number of individuals per population pool
            }}

\item{poolCol}{Character: Population pool ID. Default = \code{'POOL'}}

\item{locusCol}{Character: Locus ID. Default = \code{'LOCUS'}}

\item{refCol}{Character: Reference allele. Default = \code{'REF'}}

\item{altCol}{Character: Alternate allele. Default = \code{'ALT'}}

\item{freqCol}{Character: The reference allele frequency. Default = \code{'FREQ'}.}

\item{indsCol}{Character: The number of individuals per population pool. Default = \code{'INDS'}.}

\item{poolSub}{Character: The pools to subset out of \code{poolCol}. Default = \code{NULL}.}
}
\value{
Returns a data tablein the DADI input format.
}
\description{
Genertate DADI input from pool-seq data
}
\examples{
data(genomaliciousPi)
genomaliciousPi

dadi_inputs_pools(genomaliciousPi, poolCol='POOL', locusCol='LOCUS', refCol='REF', altCol='ALT', freqCol='P', indsCol='INDS')


}
