% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_FreqsMat.R
\docType{data}
\name{data_FreqsMat}
\alias{data_FreqsMat}
\title{An example matrix of allele frequencies}
\format{
A data table with 4 rows and 8 columns.
}
\usage{
data(data_FreqsMat)
}
\description{
A dataset containing allele frequency information for 4 populations
and 8 biallelic SNP loci.
}
\details{
Population ID are in \code{rownames(data_FreqsMat)}, and locus
ID are in \code{colnames(data_FreqsMat)}.
}
\keyword{datasets}
