% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna2codonDT.R
\name{dna2codonDT}
\alias{dna2codonDT}
\title{DNA to a condon data.table}
\usage{
dna2codonDT(cds.seq)
}
\arguments{
\item{cds.seq}{Character = A DNA coding sequence.}
}
\value{
Returns a data.table with the following columns:

\enumerate{
   \item \code{$CODON} = The codon number, 1:N.
   \item \code{$NUC} = The nucleotides comprising the codon.
   \item \code{$AMINO} = The amino acid residue.
   \item \code{$DNA} = The DNA bases.
}
}
\description{
Convert a DNA coding sequence into a data.table of codons and amino acids.
}
\details{
Assumes that the sequence is in its correct reading frame, that is,
the first DNA nucleotide is the first base of the first codon.
}
\examples{
X <- 'ATGCGTACTTCA'

dna2codonDT(X)

}
