% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna2codonDT.R
\name{dna2codonDT}
\alias{dna2codonDT}
\title{DNA to a condon data.table}
\usage{
dna2codonDT(cdsSeq, compressTab = FALSE, geneticCode = 1)
}
\arguments{
\item{cdsSeq}{Character = A DNA coding sequence.}

\item{compressTab}{Logical = Should the table be compressed? Default is TRUE,
in which case, each row is a unique codon. If FALSE, then each codon is represented
by 3 rows, one for each nucleotide comprising the codon.}

\item{geneticCode}{Integer = A value relating to the \code{numcode} argument in
\code{seqinr::translate}.}
}
\value{
Returns a data.table with the following columns:

\enumerate{
   \item \code{$CODON} = The codon number, 1:N.
   \item \code{$NUC} = The nucleotides comprising the codon.
   \item \code{$AMINO} = The amino acid residue.
   \item \code{$DNA} = The DNA bases.
}
}
\description{
Convert a DNA coding sequence into a data.table of codons and amino acids.
}
\details{
Assumes that the sequence is in its correct reading frame, that is,
the first DNA nucleotide is the first base of the first codon. Default is 1.
}
\examples{
X <- 'ATGCGTACTTCA'

dna2codonDT(X, compressTab=TRUE)

dna2codonDT(X, compressTab=FALSE)

}
