% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT2Mat_genos.R
\name{DT2Mat_genos}
\alias{DT2Mat_genos}
\title{Convert a data table of genotypes into a matrix (or vice versa)}
\usage{
DT2Mat_genos(dat, sampCol = NA, locusCol = NA, genoCol = NA,
  genoScore = "counts", flip = FALSE)
}
\arguments{
\item{dat}{Data table or Matrix: The object to transform. If this is a long data table
# of genotypes coded as per VCF specifications, e.g. ('0/0', '0/1', '1/1').
Three columns are required:
\enumerate{
   \item (1) The sampled individual ID (see param \code{sampCol}).
   \item (2) The locus ID (see param \code{locusCol}).
   \item (3) The Ref allele frequency (see param \code{genoCol}).
}
The sampled individual ID column serves as the pivot point to convert the long data table into a wide matrix.
If converting from a genotypes matrix to a data table, see argument \code{flip}.}

\item{sampCol}{Character: The column name with the sampled individual information.}

\item{locusCol}{Character: The column name with the locus information.}

\item{flip}{Logical: Instead of converting a (long) data table to a (wide) matrix,
should a (wide) matrix be converted into a (long) data table? Default = \code{FALSE}.
If \code{TRUE}, then param \code{dat} must be a matrix, with loci names as column headers,
sample IDs in the row names, and frequencies in the cells. When \code{TRUE}, params
\code{sampCol}, \code{locusCol}, and \code{genoCol} are used to structure the new matrix.}

\item{gneoCol}{Character: The column name with the genotype information.}
}
\value{
When \code{flip=FALSE}, converts a data table into a frequency matrix. When
\code{flip=TRUE}, converts a matrix into a data table with three columns: (1) \code{$SAMPLE},
the sample ID; (2) \code{$LOCUS}, the locus ID; and (3) \code{FREQ},
the Ref allele frequency.
}
\description{
Takes a data table of genotypes in long format and converts it into
a matrix in wide format (loci in columns and individuals in rows). The reverse
can also be done. See also \code{DT2Mat_freqs} for converting matrix of frequencies.
}
\examples{
data(genomaliciousPi)

# Convert a long data table to a wide matrix
freqMat <- DT2Mat_freqs(genomaliciousFreqsLong, sampCol='SAMPLE', locusCol='LOCUS', genoCol='FREQ', flip=FALSE)

# Convert a wide matrix back to a data table
freqDT <- DT2Mat_freqs(freqMat, flip=TRUE)

}
