% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_depth.R
\name{filter_depth}
\alias{filter_depth}
\title{Filter loci by sequencing depth}
\usage{
filter_depth(dat, minDP = 10, maxDP = NA, locusCol = "LOCUS", dpCol = "DP")
}
\arguments{
\item{dat}{Data table: The sequencing read information, must contain the columns:
\enumerate{
   \item The locus ID (see param \code{locusCol}).
   \item The read depth (see param \code{dpCol}).
}}

\item{minDP}{Integer: The minimum sequencing depth. Loci below this value are flagged
as 'bad' loci. Default = \code{10}. Must be specified; set \code{minDP = 0} if there is no
minimum number of reads.}

\item{maxDP}{Integer: The maximum sequencing depth. Loci above this value are flagged
as 'bad' loci. Default = \code{NA}.}

\item{locusCol}{Character: The column name with the locus information. Default = \code{'LOCUS'}.}

\item{dpCol}{Character: The column with read depth information. Default = \code{'DP'}.}
}
\value{
Returns a character vector of locus names in \code{dat$LOCUS} that conform
to the read depth values specified, i.e. 'good' loci.
}
\description{
Parses a data table of sequencing read information to determine which loci
conform to desired sequencing depth values.
}
\examples{
data(genomalicious_PoolReads)
datReads <- genomalicious_PoolReads

# Look at the sequencing coverage
datReads

# Exclude loci with coverage < 30 reads
min30 <- filter_depth(datReads, minDP=30)
min30
datReads[LOCUS \%in\% min30]

# Exclude loci with coverage < 30 and > 110 reads
min30max110 <- filter_depth(datReads, minDP=30, maxDP=110)
min30max110
datReads[LOCUS \%in\% min30max110]

# Alternatively, subset data.table to only contain the
# bad loci with coverage < 30 reads
datReads[!(LOCUS \%in\% min30)]

}
