% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_depth.R
\name{filter_depth}
\alias{filter_depth}
\title{Filter loci by sequencing depth}
\usage{
filter_depth(dat, minDP = NULL, maxDP = NULL, locusCol = "LOCUS", dpCol = "DP")
}
\arguments{
\item{dat}{Data table: The sequencing read information, must contain the columns:
\enumerate{
   \item The locus ID (see param \code{locusCol}).
   \item The read depth (see param \code{dpCol}).
}}

\item{minDP}{Integer: The minimum sequencing depth. Loci below this value are flagged
as 'bad' loci. Default is \code{NULL}.}

\item{maxDP}{Integer: The maximum sequencing depth. Loci above this value are flagged
as 'bad' loci. Default is \code{NULL}.}

\item{locusCol}{Character: The column name with the locus information. Default = \code{'LOCUS'}.}

\item{dpCol}{Character: The column with read depth information. Default = \code{'DP'}.}
}
\value{
Returns a character vector of locus names in \code{dat[[locusCol]]}
that conform to the read depth threshold (>= value of \code{minDP} and <= value
of \code{maxDP}).
}
\description{
Parses a data table of genotypes/allele frequencies and returns a list of
loci that conform to a desired read depth threshold.
}
\examples{
library(genomalicious)

data(data_4pops)

# Exclude loci with coverage < 10 reads
min10 <- filter_depth(data_4pops, minDP=10)
min10
data_4pops[LOCUS \%in\% min10]$DP \%>\% summary

# Exclude loci with coverage < 10 and > 100 reads
min10max100 <- filter_depth(data_4pops, minDP=10, maxDP=100)
min10max100
data_4pops[LOCUS \%in\% min10max100]$DP \%>\% summary

# Alternatively, subset data.table to only contain the
# bad loci with coverage < 10 reads
data_4pops[!(LOCUS \%in\% min10)]$DP \%>\% summary

}
