% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_maf.R
\name{filter_maf}
\alias{filter_maf}
\title{Filter loci by minor allele frequency (MAF)}
\usage{
filter_maf(
  dat,
  maf = 0.05,
  type = "genos",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT"
)
}
\arguments{
\item{dat}{Matrix or data.table: Default expectation is that user is supplying a matrix of allele
frequencies; i.e., loci in columns, populations in rows, and allele frequencies in cells (i.e.
\code{type=='freqs'}). \cr\cr
Alternatively, #' a data table of genotypes (i.e. \code{type=='genos'}). Loci are
expected to be biallelic. Recorded either as '/' separated alleles (0/0, 0/1 1/1),
or as counts of the Alt allele (0, 1, 2). For a data.table, expecting the columns:
\enumerate{
   \item The sample ID (see param \code{sampCol})
   \item The locus ID (see param \code{locusCol}).
   \item The genotypes (see param \code{genoCol}).
}}

\item{maf}{Numeric: The minor allele frequency. E.g. 0.05 will filter for 5%, which will remove
a locus if its frequency is < 0.05 or > 0.95. Default is 0.05, and the value
must be <=0.5.}

\item{type}{Character: Default = 'freqs', expected that \code{dat} is a matrix of allele frequencies.
Alternatively, if \code{dat} is a data table of of genotypes, set \code{type} to 'genos'.}

\item{sampCol}{Character: The column name with the sampled individual information. Default = \code{'SAMPLE'}.
Only needed when \code{dat} is a long-format data table of genotypes.}

\item{locusCol}{Character: The column name with the locus information. Default = \code{'LOCUS'}.
Only needed when \code{dat} is a long-format data table of genotypes.}

\item{genoCol}{Character: The column name with the genotype information. Default = \code{'GT'}.
Only needed when \code{dat} is a long-format data table of genotypes.}
}
\value{
Returns an integer vector of column numbers in \code{dat} that conform
to the MAF value specified. These values can then be used to filter the allele frequency matrix.
}
\description{
Parses a matrix of data table  of allele frequencies to determine which loci
conform to the desired minor allele frequency threshold.
}
\examples{
# MATRIX OF ALLELE FREQUENCIES
data(data_FreqsMat)

# Filter for MAF=0.20
mat.loci <- filter_maf(data_FreqsMat, maf=0.20, type='freqs')

data_FreqsMat[, mat.loci]

# LONG TABLE OF GENOTYPES
data(data_4pops)

# Filter for MAF=0.20
dt.loci <- filter_maf(data_4pops, maf=0.20, type='genos')

data_4pops[LOCUS \%in\% dt.loci]

}
