% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_maf.R
\name{filter_maf}
\alias{filter_maf}
\title{Filter loci by minor allele frequency (MAF)}
\usage{
filter_maf(
  dat,
  maf = 0.05,
  type = "genos",
  method = "mean",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  popCol = "POP",
  genoCol = "GT",
  freqCol = "FREQ"
)
}
\arguments{
\item{dat}{A data table of genotypes or allele frequencies. Ggenotypes are
recorded either as '/' separated alleles (0/0, 0/1 1/1), or as counts of the
Alt allele (0, 1, 2). If allele frequencies, can be either the Ref or Alt
allele, so long as it is consistent across samples, populations, loci, etc.
Expects the columns:
\enumerate{
   \item The sample ID (see param \code{sampCol}), for genotype datasets only.
   \item The locus ID (see param \code{locusCol}).
   \item The population ID (see param \code{popCol}).
   \item The genotypes (see param \code{genoCol}), or the allele frequencies
   (see param \code{freqCol})
}}

\item{maf}{Numeric: The minor allele frequency. E.g. 0.05 will filter for 5%, which will remove
a locus if its frequency is < 0.05 or > 0.95. Default is 0.05, and the value
must be <=0.5.}

\item{type}{Character: Is \code{dat} a data table of genotypes (\code{'genos'})
or a data table of allele frequencies (\code{'freqs'})? Default = \code{'genos'}.}

\item{method}{Character: The method by which MAF filtering is performed.
One of \code{'mean'}, or \code{'any_pop'}. Default = \code{'mean'}.
For \code{'mean'}, the mean MAF across populations is calculated and used to
assess the MAF threshold at each locus. For \code{'any_pop'}, if any population
has a MAF less than the threshold at a locus, then that locus will be removed.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}. Only needed when \code{type=='genos'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{popCol}{Character: The column name with population information.
Default = \code{'POP'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default = \code{'GT'}. Only needed when \code{type=='genos'}.}

\item{freqCol}{Character: The column name with the allele frequency information.
Default = \code{'freqCol'}. Only needed when \code{type=='freqs'}.}
}
\value{
Returns a character vector of locus names in \code{dat[[locusCol]]} that conform
to the MAF threshold (>= value of \code{maf}).
}
\description{
Parses a data table of genotypes/allele frequencies and returns a list of
loci that conform to a desired MAF threshold.
}
\examples{
# LONG TABLE OF GENOTYPES
data(data_Genos)

# Filter for MAF=0.20
loci.genos <- filter_maf(data_Genos, maf=0.20, type='genos')

data_Genos[LOCUS \%in\% dt.loci]

# LONG TABLE OF ALLELE FREQUENCIES
freqs_4pops <- data_Genos \%>\%
   .[, .(FREQ=sum(GT)/(length(GT)*2)), by=c('LOCUS','POP')]

loci.freqs <- filter_maf(freqs_4pops, maf=0.20, type='freqs')

}
