% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_calc.R
\name{fstat_calc}
\alias{fstat_calc}
\title{Calculate F-statistics from genotypes or allele frequencies (counts)}
\usage{
fstat_calc(
  dat,
  type,
  method,
  fstatVec = NULL,
  popCol = "POP",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT",
  countCol = "COUNTS",
  indsCol = "INDS",
  permute = FALSE,
  keepLocus = TRUE,
  numPerms = 100,
  numCores = 1
)
}
\arguments{
\item{dat}{Data table: For genotype data, a long-format data table of genotypes,
coded as '/' separated alleles ('0/0', '0/1', '1/1'). For allele frequency data,
a long-format data table of allele counts.

Columns required for \strong{both} genotypes and allele frequencies:
\enumerate{
   \item The population ID (see param \code{popCol}).
   \item The locus ID (see param \code{locusCol}).
}
Columns required only for \strong{genotypes}:
\enumerate{
   \item The sample ID (see param \code{sampCol}).
   \item The genotypes (see param \code{genoCol}).
}
Columns required only for \strong{allele frequencies}:
\enumerate{
   \item The allelic count column (see param \code{countCol}).
   \item The number of individuals used to obtain the allele frequency
   estimate (see param \code{indsCol}).
}}

\item{type}{Character: One of \code{'genos'} or \code{'freqs'}, to calculate
F-statistics from genotype or allele frequency data, respectively.}

\item{method}{Character: One of \code{'global'} or \code{'pairwise'} for
global or pairwise F-statistics, respectively.}

\item{fstatVec}{Character: A vector of F-statistics to calculate. This is only
applicable for genotype data, \code{type=='genos'}. Must include one of \code{"FST"},
\code{"FIS"}, or \code{"FIT"}.}

\item{popCol}{Character: The column name with the population information.
Default is \code{'POP'}.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default is \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default is \code{'GT'}.}

\item{countCol}{Character: The column name with the allele count information.
Default is \code{'FREQ'}. Counts for each allele need to be separated with a
comma, starting with the Ref allele, followed by each subsequent Alt allele.
E.g., '0,25', or '5,7,10', for a locus with 2 alleles and 3 alleles, respectively.
You must code alleles within a locus at same positions in the character
string across all populations.}

\item{indsCol}{Character: The column name with the number of individuals
contributing to the allele freuqency estimate. Default is \code{indsCol}.}

\item{permute}{Logical: Should permutations be performed to test the
statistical significance of F-statistics? Default is \code{FALSE}.
Can only be performed on genotype data, \code{type=='genos'}.}

\item{keepLocus}{Logical: Should locus-specific estimates of F-statistics be
kept? Default is TRUE. Dropping locus-specific estimates will dramatically
save memory and the size of the returned list.}

\item{numPerms}{Integer: The number of permutations to perform. Default is 100.}

\item{numCores}{Integer: The number of cores to use when running permutations. Default is 1.}
}
\value{
A list is returned with three indexes.

The first index is \code{$genome}, the genome-wide F-statistics. If global
estimates were requested, \code{global==TRUE}, then this is just a single row;
the estimates across all populations. If pairwise esimates were requested,
\code{pairwise==TRUE}, then there are \code{$POP1} and \code{$POP2}, which
represent two populations tested.

The second index is \code{$locus}, the locus-specific F-statistics. This is
a data table with a \code{$LOCUS} column for global estimates at each locus.
when \code{global==TRUE}. If pairwise population estimates have been requested,
\code{pairwise==TRUE}, then there are \code{$POP1} and \code{$POP2}, which
represent the two populations tested.

The third index is \code{$permute}, the permutation results. This index will
be \code{NULL} when frequencies are used, i.e., \code{type=='freqs'}, and
will only contain data if \code{type=='genos'} and \code{permute==TRUE}.
\code{$permute} is itself a list, with two subindexes:
\enumerate{
   \item \code{$fstat}: The permuted F-statistics. If \code{global==TRUE}, then
   this will simply be a single row of global estimates. If \code{pairwise==TRUE},
   then this will be a data table with columns \code{$POP1}, \code{$POP2}, and
   a column for each F-statistic.

   \item \code{$pval}: The permuted p-values. This is a long-format data table.
   If \code{global==TRUE}, then there are two column: \code{$STAT}, which
   contains the F-statistic; and \code{$PVAL}, which contains the global
   permuted p-value. If \code{pairwise==TRUE}, then there will two additional
   columns, \code{$POP1} and \code{$POP2}.
}
}
\description{
Function takes a genotypes or allele frequencies in a long-format data table
and calculates Weir & Cockerham's F-statistics (Weir & Cockerham, 1984).
Permutations can be used to test statistical significance of F-statistics in
genotype data sets. Can deal with multiallelic data. See Details for more information.
}
\details{
With genotype data, the F-statistics FST, FIS, and FIT can be calculated.
Only FST can be calculated from allele frequency data.

F-statistics from genotype data are calculated from the variance components
'a', 'b', and 'c', which have been standardised for observed heterozygosity.
FST from allele frequency data uses an estimate of the expected heterozygosity.

Permutation tests for genotype data involve random shuffling of individuals
among populations, recalculating F-statistics, and testing the hypothesis
that the permuted F-statistic > observed F-statistic. The p-value represents the
proportion of permutation that were TRUE to this expression. That is,
if no permuted values are greater than the observed, p=0. Likewise, if all
the permuted values are greater than the observed, p=1.
}
\examples{
library(genomalicious)

data(data_Genos)
data(data_PoolFreqs)
data(data_PoolInfo)

# Set genotypes as characters
data_Genos$GT \%>\% head
data_Genos[, GT:=genoscore_converter(GT)]
data_Genos$GT \%>\% head

# Set allele counts and individuals in pool-seq data
data_PoolFreqs \%>\% head
data_PoolInfo \%>\% head

data_PoolFreqs[, COUNTS:=paste(RO,AO,sep=',')]

data_PoolFreqs$INDS <- data_PoolInfo$INDS[
match(data_PoolFreqs$POOL, data_PoolInfo$POOL)
]

head(data_PoolFreqs)

# Genotypes and global F-statistics
geno_global_f <- fstat_calc(
dat=data_Genos,
type='genos', method='global', fstatVec=c('FST','FIS','FIT'),
popCol='POP', sampCol='SAMPLE',
locusCol='LOCUS', genoCol='GT',
permute=FALSE
)

# Genotypes and pairwise F-statistics
geno_pair_f <- fstat_calc(
dat=data_Genos,
type='genos', method='pairwise', fstatVec=c('FST','FIS','FIT'),
popCol='POP', sampCol='SAMPLE',
locusCol='LOCUS', genoCol='GT',
permute=FALSE
)

# Allele frequencies (from counts) and global FST
freqs_global_f <- fstat_calc(
dat=data_PoolFreqs,
type='freqs', method='global', fstatVec=NULL,
popCol='POP', locusCol='LOCUS',
countCol='COUNTS', indsCol='INDS',
permute=FALSE
)

# Allele frequencies (from counts) and pairwise FST
freqs_pair_f <- fstat_calc(
dat=data_PoolFreqs,
type='freqs', method='pairwise', fstatVec=NULL,
popCol='POP', locusCol='LOCUS',
countCol='COUNTS', indsCol='INDS',
permute=FALSE
)

}
\references{
Weir & Cockerham (1984) Evolution. DOI: 10.1111/j.1558-5646.1984.tb05657.x
Weir et al. (2002) Annals of Human Genetics. DOI: 10.1146/annurev.genet.36
}
