% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstWC_genos.R
\name{fstWC_genos}
\alias{fstWC_genos}
\title{Calculate Weir and Cockerham's FST from genotypes}
\usage{
fstWC_genos(dat, popCol = "POP", sampCol = "SAMPLE", locusCol = "LOCUS",
  genoCol = "GT", permTest = FALSE, bootCI = FALSE, iters = 100,
  doPairs = FALSE, doDist = FALSE, perLocus = FALSE)
}
\arguments{
\item{dat}{Data table: A long format data table of biallelic genotypes,
coded as '/' separated alleles ('0/0', '0/1', '1/1'), or counts
of the Alt alleles (0, 1, 2, repsectively).
Three columns are required:
\enumerate{
   \item The population ID (see param \code{popCol}).
   \item The sampled individual ID (see param \code{sampCol}).
   \item The locus ID (see param \code{locusCol}).
   \item The genotype (see param \code{genoCol}).
}}

\item{popCol}{Character: The column name with the population information.
Default = \code{'POP'}.}

\item{sampCol}{Character: The column name with the sampled individual information.
Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default = \code{'GT'}.}

\item{permTest}{Logical: Should a permutation test for significance be
conducted? Default = \code{FALSE}, see Details.}

\item{bootCI}{Logical: Should bootstrap resampling be conducted to estimate
the 2.5% and 97.5% confidence intervals? Default = \code{FALSE}, see Details.}

\item{iters}{Integer: The number of permutations or bootstraps to conduct.}

\item{doPairs}{Logical: Should pairwise FSTwc be calculated (TRUE) or the
among population FSTwc (FALSE)?}

\item{doDist}{Logical: Should a a distance matrix of FST be returned as well?
Will only run if \code{doPairs==TRUE}.}

\item{perLocus}{Logical: Should the per locus FST be returned? Default is FALSE.}
}
\value{
A list, contents depend on whether analysis is among populations
or pairwise. \cr
For among populations:
\enumerate{
   \item \code{$fst.mean}: Numeric, the mean FST across loci
   \item \code{$fst.locus}: Numeric, the locus-specific FST values
   \item \code{$fst.perm.pval}: Numeric, the p-value from permutation tests,
   the number of permutations that produced FST > observed FST.
   \item \code{$fst.boot.ci}: Numeric, the bootstrap 2.5% and 97.5% percentiles.
}
For pairiwse populations:
\enumerate{
   \item \code{$fst.mean}: Data table, contains the columns \code{$POP1},
   \code{$POP2}, \code{$FST}. Will also contain \code{$PVAL} for
   permutation tests, and \code{$CI.025} and \code{$CI.975} for
   bootstrap confidence intervals.
   \item \code{$fst.locus}: Data table, contains the columns \code{$POP1},
   \code{$POP2}, \code{$LOCUS},and \code{$FST}.
   \item \code{$fst.dist}: Dist, a distance matrix of mean FST values.
}
}
\description{
Takes a long format data table of biallelic genotypes and calculates
Weir and Cockerham's FST, i.e. theta (Weir & Cockerham, 1984).
}
\details{
Significance testing using permutations (\code{permTest==TRUE})
randomly assigns individuals to populations before calculating FST.
The p-value from this test reflects the proportion of
permuted FST values that were greater than the observed FST. NOTE: when
observed FST > 0, this tends to always give a significant result because
Weir & Cockerham's FST always returns negative FST values when individuals
are fully randomised among populations (based on personal trails). Sometimes
observed FST < 0 can also be significant, but this is because randomisation
can create less negative FST values. \cr

Estimating confidence intervals using bootstrap resampling
(\code{bootCI==TRUE}) can provide insight into how likely the FST
estimate overlaps with zero, and might be a more conservative test
for testing statistical significance.
}
\examples{
data(genomalicious_4pops)

# Among popuation FST, with permutation tests
# and bootstrap confidence intervals.
fst_among <- fstWC_genos(genomalicious_4pops, permTest=TRUE
             , bootCI=TRUE, iters=50)

# Pairwise FST, with per locus estimates
# and a distance matrix
fst_pairs <- fstWC_genos(genomalicious_4pops, doPairs=TRUE
             , doDist=TRUE, perLocus=TRUE)

}
