% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genos2freqs.R
\name{genos2freqs}
\alias{genos2freqs}
\title{Generate a matrix of allele frequencies from genotypes}
\usage{
genos2freqs(
  dat,
  popCol = "POP",
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT",
  returnMat = TRUE
)
}
\arguments{
\item{dat}{Data table: It is expected that there are only two alleles, and therefore, only three possible genotypes:
0/0, 0/1 (or 1/0), and 1/1, where the Ref allele is '0'. This data.table needs the following columns:
\enumerate{
   \item The population ID (see param \code{popCol}).
   \item The sample ID (see param \code{sampCol})
   \item The locus ID (see param \code{locusCol}).
   \item The genotypes (see param \code{genoCol}).
}}

\item{popCol}{Character: The column name with the sampled individual information. Default = \code{'POP'}.}

\item{sampCol}{Character: The column name with the sampled individual information. Default = \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information. Default = \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information. Default = \code{'GT'}.}

\item{returnMat}{Logical: Should an allele frequency matrix be returned?
Default = \code{TRUE}, but if \code{FALSE}, will return a data table.}
}
\value{
Returns a matrix of allele frequencies for the Alt allele, or
a data table with the columns \code{$POP}, \code{$SAMPLE}, \code{$LOCUS}
and \code{$ALT.FREQ}.
}
\description{
Parses a data table of genotypes and returns a matrix of allele frequencies.
}
\examples{
# Import genotype data
data(genomalicious_4pops)
genomalicious_4pops

# Convert to frequency matrix
genos2freqs(genomalicious_4pops)

}
