% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/He_freqs.R
\name{He_freqs}
\alias{He_freqs}
\title{Calculate expected heterozygosity from allele frequencies}
\usage{
He_freqs(dat, popCol = "POP", locusCol = "LOCUS", freqCol = "FREQ")
}
\arguments{
\item{dat}{Matrix or Data table/frame: Can take a matrix of allele frequencies (row names
are populations, column names are loci, cells are frequencies of Ref alleles) or
a data table/frame in long format (requires)}

\item{popCol}{Character: The column name in the long format data table/frame \code{dat}
corresponding to the population ID.}

\item{locusCol}{Character: The column name in the long format data table/frame \code{dat}
corresponding to the locus ID.}

\item{freqCol}{Character: The column name in the long format data table/frame \code{dat}
corresponding to the frequency of the Ref allele for a locus in a population..}
}
\value{
Returns a list with the following indexed items (data tables): \cr
\enumerate{
    \item \code{popXlocus} = He for each population x locus combination (columns:
    \code{$POP}, \code{$LOCUS}, \code{$HE}).
    \item \code{popAv} = Average He per populations (columns: \code{$POP}, \code{$HE.AV}).
    \item \code{locusAv} = Average He per locus (columns: \code{$LOCUS}, \code{$HE.AV})
}
}
\description{
Takes a matrix or data table/frame of allele frequencies for different populations,
at different loci, and returns various statistics for expected heterozygosity (He).
Assumes biallelic data.
}
\details{
Calculates He as 2 * p * q; where p = the Ref allele frequency, and
q = the Alt allele frequency.
}
\examples{
# Data as matrix
data(genomalicious_Freqs)
He1 <- He_freqs(genomalicious_Freqs)

# Data as long data table
data(genomalicious_PoolPi)
He2 <- He_freqs(genomalicious_PoolPi, popCol='POOL', locusCol='LOCUS', freqCol='PI')

}
