% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_calc.R
\name{het_calc}
\alias{het_calc}
\title{Calculate heterozygosity from genotypes or allele frequencies}
\usage{
het_calc(
  snpData,
  extraData,
  method,
  type,
  chromCol = "CHROM",
  posCol = "POS",
  locusCol = "LOCUS",
  sampCol = "SAMPLE",
  genoCol = "GT",
  popCol = "POP",
  roCol = "RO",
  aoCol = "AO",
  covCol = "COV.SITES",
  indsCol = "INDS"
)
}
\arguments{
\item{snpData}{Data table: Genotypes for individuals or frquencies (read coiunts)
for populations in long-format. See Details for parameterisation.}

\item{extraData}{Data table: Extra info for individual or populations.
See Details for parameterisation.}

\item{method}{Charachter: One of \code{'genomic'} or \code{'snpwise'} to
calculate genomic or SNP-wise heterozygosity, respectively.}

\item{type}{Character: One of \code{'genos'} or \code{'freqs'}, to calculate
heterozygosity on genotype or frequencies, respectively.}

\item{chromCol}{Character: The column with chromosome ID.
Default is \code{'CHROM'}.}

\item{posCol}{Character: The column with the positional info.
Default is \code{'POS'}.}

\item{locusCol}{Character: The column with locus ID.
Default is \code{'LOCUS'}.}

\item{sampCol}{Character: The column name with the sampled individual ID.
Default is \code{'SAMPLE'}.}

\item{genoCol}{Character: The column with the genotype info.
Default is \code{'GT'}. Genotypes should be scored as alleles separated by
'/', e.g., '0/0', '0/1', '1/1', etc.}

\item{popCol}{Character: The column name with population ID.
Default is \code{'POP'}.}

\item{roCol}{Character: The column with reference allele counts.
Default is \code{'RO'}.}

\item{aoCol}{Character: The column with alternate allele counts.
Default is \code{'AO'}.}

\item{covCol}{Character: The column with the number of genomic sites covered
per chromosome. Default is \code{'COV.SITES'}.}

\item{indsCol}{Character: The column name with the number of pooled individuals
per population per chromosome. Default is \code{'INDS'}.}
}
\value{
Returns a data table of heterozygosity estimates per sample or population.
Genomic heterozygosity is reported per chromosome, SNP-wise heterozygosity is
reported across all SNPs.
}
\description{
This function takes in a long format data table of genotypes or allele
frequencies (allelic counts) and calculates heterozygosity. Heterozygosity
is calculated as either the SNP-wise heterozygosity (at each SNP, excluding
monomorphic sites), or as the genomic heterozygosity (the SNP-wise
heterozygosity standardised for total sites assayed, including monomorphic
sites). For indivdiual genotypes, you can use multiallelic data.
For population allele frequencies, you must code as biallelic reference and
alternate alleles.
}
\details{
The genomic heterozygosity (also known as the autosomal
heterozygosity) has been demonstrated to be the more accurate and robust
measure of heterozygosity (Schmidt et al. 2021). SNP-wise heterozygosity can
suffer from sampling biases (filtering, missing data, sample size, etc).
Note, estimates of genomic heterozygosity are orders of magnitude less than
SNP-wise heterozygosity (so do not be alarmed if you see very small values!).

Heterozygsity for population pools is calculated using the method from
Ferretti et al. (2013). You can calculate genomic heterozygosity for population
pools (standardising by total covered sites), or SNP-wise heterozygosity too
(standardising by the number of polymorphic sites observed).

You must specify both \code{type} and \code{method}. This will dicate the
required column needed for \code{snpData} and \code{extraData}.

If \code{type=='genos'} and \code{method=='genomic'}:
\enumerate{
   \item \code{snpData} requires columns specified in:
   \code{chromCol}, \code{posCol}, \code{locusCol} and \code{sampCol}

   \item \code{extraData} requires columns specified in:
   \code{chromCol}, \code{sampCol}, and \code{covCol}.
}

If \code{type=='genos'} and \code{method=='snpwise'}
\enumerate{
   \item \code{snpData} requires columns specified in:
   \code{chromCol}, \code{posCol}, \code{locusCol} and \code{sampCol}

   \item \code{extraData} will NOT be used.
}

If \code{type=='freqs'} and \code{method=='genomic'}:
\enumerate{
   \item \code{snpData} requires columns specified in:
   \code{chromCol}, \code{posCol}, \code{locusCol}, \code{popCol},
   \code{roCol}, and \code{aoCol}.

   \item \code{extraData} requires columns specified in:
   \code{chromCol}, \code{sampCol}, \code{covCol}, and \code{indsCol}.
}

If \code{type=='freqs'} and \code{method=='snpwise'}:
\enumerate{
   \item \code{snpData} requires columns specified in:
   \code{chromCol}, \code{posCol}, \code{locusCol}, \code{popCol},
   \code{roCol}, and \code{aoCol}.

   \item \code{extraData} requires columns specified in:
   \code{chromCol}, \code{sampCol}, and \code{indsCol}.
}
}
\examples{
library(genomalicious)

data(data_Genos)
data(data_PoolFreqs)

# Convert genos to characters
data_Genos[, GT:=genoscore_converter(GT)]

# Make extra data for the samples and populatin pools
extraSampInfo <- CJ(
  SAMPLE=unique(data_Genos$SAMPLE),
  CHROM=unique(data_Genos$CHROM),
  COV.SITES=150
  )

extraPoolInfo <- CJ(
  POP=unique(data_PoolFreqs$POP),
  CHROM=unique(data_PoolFreqs$CHROM),
  COV.SITES=150,
  INDS=30
  )

# Genomic heterozygosity of individuals, per chromosome/contig
het_calc(data_Genos, extraSampInfo, type='genos', method='genomic')

# SNP-wise heterozygosity of indivdiuals, SNP-wise
het_calc(data_Genos, extraSampInfo, type='genos', method='snpwise')

# Genomic heterozygosity of population pools, per chromosome/contig
het_calc(data_PoolFreqs, extraPoolInfo, type='freqs', method='genomic')

# Genomic heterozygosity of population pools, SNP-wise
het_calc(data_PoolFreqs, extraPoolInfo, type='freqs', method='snpwise')
}
\references{
Ferretti et al. (2013) Molecular Ecology. DOI: 10.1111/mec.12522 \cr
Schmidt et al. (2021) Methods in Ecology and Evolution. DOI: 10.1111/2041-210X.13659
}
