% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idip_DTfreqs.R
\name{idip_DTfreqs}
\alias{idip_DTfreqs}
\title{Calculate diversity statistics with \code{IDIP} from a data table of frequencies}
\usage{
idip_DTfreqs(freqDat, strucMat, popCol = "POP", locusCol = "LOCUS",
  freqCol = "FREQ", alleleCol = NA)
}
\arguments{
\item{freqDat}{Data table: a long-format data table with SNP genotypes
allele frequencies. Three columns are required:
\enumerate{
   \item (1) The population ID (see param \code{popCol}).
   \item (2) The locus ID (see param \code{locusCol}).
   \item (3) The allele frequencies (see param \code{freqCol}).
}}

\item{strucMat}{Matrix: Hierarchy of "aggregates", minimum of 2 rows and
as many columns as the lowest units of the hierarchy. 1st row is the top most
level (e.g. ecosystem or metapopulation), with lower levels in proceding rows.}

\item{popCol}{Character: The column name with the population information, i.e.
the lowest levels of the population hierarchy (last row of \code{strucMat}).
Default = \code{'POP'}.}

\item{locusCol}{Character: The column name with the locus information.
Default = \code{'LOCUS'}.}

\item{freqCol}{Character: The column name with the allele frequency information.
Default = \code{'FREQ'}.}

\item{alleleCol}{Character: If number of alleles > 2, then specify a column that
contains the allele information, otherwise, Default = \code{NA}. Ideally,
all alleles possible at a locus should recorded for each population.}
}
\value{
A matrix with columns representing loci, rows are the statistics
from the \code{IDIP} analysis.
}
\description{
Calculate diversity statistics with \code{IDIP} from a data table of frequencies
}
\examples{
data(genomalicious_PoolPi)
freqDat <- genomalicious_PoolPi
strucMat <- matrix(c(rep('metapop', 4)
               , paste('Group', c(1,1,2,2))
               , paste('Pop', 1:4, sep='')), ncol=4, byrow=TRUE)

metaPop <- matrix(c(rep('metapop', 4)
               , paste('Group', c(1,1,2,2))
               , paste('Pop', 1:4, sep='')), ncol=4, byrow=TRUE)

idip4pops <- idip_DTfreqs(freqDat=genomalicious_PoolPi, strucMat=metaPop
             , popCol='POOL', locusCol='LOCUS', freqCol='PI')

# Simple example using 3 alleles
als3 <- data.table(POP=c(rep('Pop1', 3), rep('Pop2', 3))
           , LOCUS=rep('Locus1', 6), ALLELE=c(letters[1:3], letters[1:3])
           , FREQ=c(0.5, 0.25, 0.25, 0.9, 0.05, 0.05))
als3

idip3als <- idip_DTfreqs(freqDat=als3
               , strucMat=matrix(c(rep('metapop', 2), 'Pop1', 'Pop2'), ncol=2, byrow=TRUE)
               , popCol='POP'
               , locusCol='LOCUS'
               , freqCol='FREQ'
               , alleleCol='ALLELE')

}
