% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_overlap.R
\name{locus_overlap}
\alias{locus_overlap}
\title{Proability of random null overlap among groups for outlier loci}
\usage{
locus_overlap(lociList, lociSig, perms = 1000)
}
\arguments{
\item{lociList}{List: A list of loci analysed for each group. Each index is a group,
containing a vector loci names.}

\item{lociSig}{List: A list of significant loci for each group. Each index is a group,
containing a vector of loci names.}

\item{perms}{Integer: The number of permutations to run for significance testing. Default = 1000}
}
\value{
The value returned is the proportion of null permutations that were greater
than the observed value (i.e. the empirical number of significant loci shared between groups).
}
\description{
This functions calculates the null probability of observing
overlap in outlier loci between 2+ groups with respect to a scenario
whereby loci are drawn at random from a set of loci.
}
\details{
This non-parametric sytle test is not a strict population genetic statistical
method per se, but provides an informal way to compare parameters of an
outlier analysis (number of groups, number of analysed loci,
and the number of observed outlier loci) to random chance.


It is very important that the order of indices in \code{lociList} and \code{lociSig}
are in the same, because each index is expected to correpond to a group.

The hypothesis being tested is that the number of observed loci overlapping
among groups is greater than that observed due to random chance.

It is assumed that the amount of overlap is across all groups. E.g. if only 2 groups
are provided, then a locus must be significant in both groups. If 3 groups are provided,
then a locus must be significant in all 3 groups.

Permutations are run where loci are randomly drawn from each group's set of loci at a size equivalent
to that deemed statistically significant in empircal tests. In other words, if a test
determined 2 loci out of 100 were significant, a null draw would randomly sample 2 loci from the set of 100.
}
\examples{
grps <- list(A=c(1:101), B=c(20:100), C=c(1:50, 70:120))

sig <- list(A=c(1,8,88), B=c(88, 101), C=c(88, 102, 118, 120))

locus_overlap(lociList=grps, lociSig=sig, perms=1000)

}
