% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2DT.R
\name{pca2DT}
\alias{pca2DT}
\title{Convert a \code{prcomp} object to a long-format \code{data.table}}
\usage{
pca2DT(pcaObj)
}
\arguments{
\item{pcaObj}{Prcomp: An object produced by \code{prcomp}. If produced from
\code{genomalicious::pca_genos}, it may contain a \code{$pops} indexed
item, which contains the population designations for each sample in
\code{pcaObj$x}.}
}
\value{
Returns a long-format data.table with columns \code{$POP}, the
population IDs, \code{$SAMPLE}, the sample IDs, \code{$AXIS}, the PC axis ID,
and \code{$SCORE}, the score on the PC axis.
}
\description{
Convert a \code{prcomp} object to a long-format \code{data.table}
}
\examples{
data(data_4pops)

PCA <- pca_genos(data_4pops, popCol='POP')

pca2DT(PCA)

}
