% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2DT.R
\name{pca2DT}
\alias{pca2DT}
\title{Convert a \code{prcomp} object to a long-format \code{data.table}}
\usage{
pca2DT(pcaObj, pops = NULL, subAxes = NULL)
}
\arguments{
\item{pcaObj}{Prcomp: An object produced by \code{prcomp}. If produced from
\code{genomalicious::pca_genos}, it may contain a \code{$pops} indexed
item, which contains the population designations for each sample in
\code{pcaObj$x}.}

\item{pops}{Character: An optional vector of population IDs. Default is \code{NULL}.
Be sure to check that they match samples in rows of \code{pcaObj$x}.}

\item{subAxes}{Integer: An optional vector of values to subset the PC axes.
Default is NULL.}
}
\value{
Returns a long-format data.table with columns \code{$POP}, the
population IDs (if data available), \code{$SAMPLE}, the sample IDs,
\code{$PC[x]}, the individual PC axes, with \code{[x]} denoting the axis number.
}
\description{
Convert a \code{prcomp} object to a long-format \code{data.table}
}
\examples{
library(genomalicious)
data(data_4pops)

# With populations
PCA.pops <- pca_genos(data_4pops, popCol='POP')

pca2DT(PCA.pops) \%>\% print

# Without populations and all axes (default)
PCA.nopops <- pca_genos(data_4pops)

pca2DT(PCA.nopops) \%>\% print

# Manual populations, with just the first three axes
pops <- rownames(PCA.nopops$x) \%>\%
   sub('Ind', '', .) \%>\%
   sub('_.*', '', .)

pops

pca2DT(PCA.nopops, pops=pops, subAxes=1:3) \%>\% print

}
