% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Plot PCA results}
\usage{
pca_plot(dat, type = "scatter", axisIndex = c(1, 2), pops = NULL,
  popColours = NULL, look = "ggplot")
}
\arguments{
\item{dat}{Prcomp/Data table/Data frame/Matrix or Numeric: The class of this
object depends on what plot you want to make, as specified by \code{type}.
If making a...
\itemize{
\item \strong{Scatter plot}: \code{dat} contains the PCA scores for
each individual. If not using a \code{prcomp} object, the tabular data must
be wide-format: rows are indivdiuals, columns are PC axes.
\item \strong{Scree plot} of eigenvalues, or plotting the \strong{cumulative
explained variance}: \code{dat} can be a \code{prcomp} object, or simply
a numeric vector containing the eigenvalues of the PCA.
}}

\item{type}{Character: What type of plot to make: scatter (\code{'scatter'})
the scree plot of eigenvalues (\code{'scree'}), or the cumulative explained
variance (\code{'expvar'}).}

\item{axisIndex}{Integer: Vector of lenght = 2. Corresponds to the column
index to plots, i.e. the PC axes. Default = \code{c(1,2)}, the first and second
PC axis column tabulated in \code{dat}. Only valid when \code{type=='scatter'}.}

\item{pops}{Character: A vector of population ID, should match the
rows in \code{dat}, but is an optional argument. Default = \code{NULL}.
If \code{dat} is a \code{prcomp} object, function will search for \code{dat$pops}
to assign to this argument. Only valid when \code{type=='scatter'}.}

\item{popColours}{Character: A vector of colours to use for each unique population
in \code{pops}, but is an optional argument. Default = \code{NULL}.
The name of each colour must correspond to a population in \code{pops}.
Only valid when \code{type=='scatter'}.}

\item{look}{Character: The look of the plot. Default = \code{'ggplot'}, the
typical gray background with gridlines produced by \code{ggplot2}. Alternatively,
when set to \code{'classic'}, produces a base R style plot.}
}
\value{
Prints the plot to screen, and also returns an object of
\code{gg}/\code{ggplot} class.
}
\description{
Plots different results of a PCA, e.g.: scatter, scree, and cumulative
explained vairance plots. See also \code{pca_DTinds()} to perform PCA on SNP genotypes.
}
\examples{
# Data
data(genomalicious_4pops)
datGt <- genomalicious_4pops

# Conduct the PCA with Patterson et al.'s (2006) normalisation, and
# population specified
pca <- pca_DTgenos(dat=datGt, scaling='patterson', popCol='POP')

# Plot the PCA
pca_plot(pca)

# Get more specific on scatter
pca_plot(pca
            , axisIndex=c(2,3)
            , pops=pca$pops
            , popColours=c(Pop1='gray30', Pop2='royalblue', Pop3='palevioletred3', Pop4='plum2')
            , look='classic')

# Plot scree of eigenvalues, using prcomp or numeric object
pca_plot(pca, type='scree')
pca_plot(pca$sdev^2, type='scree')

# Plot cumulative explained variance, using prcomp or numeric object
pca_plot(pca, type='expvar', look='classic')
pca_plot(pca$sdev^2, type='expvar', look='classic')

}
