% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolfstat_DT2pooldata.R
\name{poolfstat_DT2pooldata}
\alias{poolfstat_DT2pooldata}
\title{Calculate FST with \code{poolfstat} from a data table of read counts}
\usage{
poolfstat_DT2pooldata(dat, pool.info)
}
\arguments{
\item{dat}{Data table: Contains read counts, e.g. like that been
produced by the function \code{vcf2DT}. Must contain all the following columns:
\enumerate{
   \item \code{$CHROM} The chromosome (contig) ID.
   \item \code{$POS} The variant position on the chromosome.
   \item \code{$REF} The reference allele.
   \item \code{$ALT} The alternate allele.
   \item \code{$LOCUS} The locus ID.
   \item \code{$POOL} The pool ID.
   \item \code{$AO} The number of reads supporting the alternate allele.
   \item \code{$RO} The number of reads supporting the reference allele.
}}

\item{pool.info}{Data table: Contains the sample sample sizes (number of diploids) for
for each unique pool listed in \code{dat$POOL}. Requires two columns:
\enumerate{
   \item \code{$POOL} The pools listed in \code{dat$POOL}.
   \item \code{$INDS} The number of diploid individuals for the pools.
}}
}
\value{
Returns a \code{pooldata} object as per the \code{poolfstat} package.
}
\description{
Takes a data table of read counts and creates an object of class
\code{poolfstat}. The FST for the pools in the data table is calculated using
the function \code{poolfstat::computeFST}. Also requires pool size information.
}
\examples{
library(genomalicious)

# Load in the pool metadata and reads
data(data_PoolInfo)
data(data_PoolFreqs)

# Pool info
data_PoolInfo

# Pool reads in $DP, $AO, and $RO
data_PoolFreqs[, c('POOL','DP','AO','RO')]

# Make pooldata object
pooldataObj <-poolfstat_DT2pooldata(data_PoolFreqs, data_PoolInfo)

# And go back to data table
pooldataTab <- poolfstat_pooldata2DT(pooldataObj)

}
