% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_miss_genos.R
\name{replace_miss_genos}
\alias{replace_miss_genos}
\title{Replace missing genotypes}
\usage{
replace_miss_genos(
  dat,
  sampCol = "SAMPLE",
  locusCol = "LOCUS",
  genoCol = "GT",
  popCol = NULL,
  methodReplace = "mode"
)
}
\arguments{
\item{dat}{Data table: A long data table, e.g. like that imported from
\code{vcf2DT}. Genotypes can be coded as '/' separated characters
(e.g. '0/0', '0/1', '1/1'), or integers as Alt allele counts (e.g. 0, 1, 2).
Must contain the following columns,
\enumerate{
  \item The sampled individuals (see param \code{sampCol}).
  \item The locus ID (see param \code{locusCol}).
  \item The genotype column (see param \code{genoCol}).
}}

\item{sampCol}{Character: The column name with the sampled individual information.
Default is \code{'SAMPLE'}.}

\item{locusCol}{Character: The column name with the locus information.
Default is \code{'LOCUS'}.}

\item{genoCol}{Character: The column name with the genotype information.
Default is \code{'GT'}.}

\item{popCol}{Character: An optional argument. The column name with the
population information. Default is \code{NULL}. If specified, genotype
replacement at each locus is done per population, not across all
sampled individuals.}

\item{methodReplace}{Character: The way in which missing genotypes
are replaced. Either by the code{'median'} (middle) genotype, or by
the \code{'mode'} (most common) genotype. Default = \code{'mode'}.}
}
\description{
For each locus, missing genotypes are replaced with either
the median or mode genotype. Can be done across all sampled individuals
or by population. Loci must be biallelic. NOTE: it is recommended
that missing genotypes are imputed, but if you need a
quick-and-dirty approach, this function might be useful for
preliminary analyses, or if missing data is very low.
}
\details{
If genotypes are coded as characters, \code{NA} or \code{'./.'}
should be used to code missing genotypes. Otherwise if genotypes
are coded as integers, \code{NA} should code missing genotypes.
}
\examples{
# Create a link to raw external datasets in genomalicious
genomaliciousExtData <- paste0(find.package('genomalicious'), '/inst/extdata')

# Use this to create a path to the genomalicious ind-seq VCF
vcfPath <- paste0(genomaliciousExtData, '/genomalicious_indseq.vcf')

# Import the VCF
indseq <- vcf2DT(vcfPath)
indseq

# Add a population column
indseq[, POP:=sub(
        pattern='Ind'
        , replacement=''
        , x=strsplit(SAMPLE, '.', fixed=TRUE)[[1]][1])
    , by=c('LOCUS', 'SAMPLE') ]

# Sites with missing data
genomiss <- which(indseq$GT=='./.')
genomiss
indseq[genomiss,]

# Replace genotypes using the mode across all individuals
indseq_mode_all <- replace_miss_genos(
    dat=indseq
    , sampCol='SAMPLE'
    , locusCol='LOCUS'
    , popCol=NULL
    , methodReplace='mode')

# Replace genotypes using the mode within each population
indseq_mode_pops <- replace_miss_genos(
    dat=indseq
    , sampCol='SAMPLE'
    , locusCol='LOCUS'
    , popCol='POP'
    , methodReplace='mode')

# Replace genotypes using the median within each population
indseq_med_pops <- replace_miss_genos(
    dat=indseq
    , sampCol='SAMPLE'
    , locusCol='LOCUS'
    , popCol='POP'
    , methodReplace='median')

# Compare between methods
cbind(indseq_mode_all[genomiss, GT]
    , indseq_mode_pops[genomiss, GT]
    , indseq_med_pops[genomiss, GT])

}
